% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_alpha.R
\name{add_alpha}
\alias{add_alpha}
\title{add_alpha}
\usage{
add_alpha(col, alpha = 1)
}
\arguments{
\item{col}{Color}

\item{alpha}{Alpha level (between 0 and 1)}
}
\value{
Color with alpha level applied
}
\description{
A function to add alpha levels to RGB colors
}
