% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_in.R
\name{add_categorical_variable_from_raw_input}
\alias{add_categorical_variable_from_raw_input}
\title{add_categorical_variable_from_raw_input}
\usage{
add_categorical_variable_from_raw_input(
  name_in = "",
  acronym = "",
  variable_description = "",
  levels_in = c(),
  levels_out = c(),
  df_in = IN
)
}
\arguments{
\item{name_in}{Nome lungo (descrizione) della variabile nel data-frame di input.}

\item{acronym}{Nome (sigla) della variabile nel data-frame di input.}

\item{variable_description}{Nome lungo o descrizione della variabile.}

\item{levels_in}{Vettore con i livelli nella variabile di input.}

\item{levels_out}{Vettore con i livelli per la variabile di output.}

\item{df_in}{Data-frame di inpunt.}

\item{value}{Valore numerico.}

\item{decimals}{Numero di decimali da mostrare.}
}
\value{
Un vettore (utile per inserirlo in un data-frame).
}
\description{
Funzione per ottenere una variabile categorica a partire da un input di inserimento.
}
