% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{arrotonda}
\alias{arrotonda}
\title{Arrotonda}
\usage{
arrotonda(value, decimals = 3)
}
\arguments{
\item{value}{Valore numerico.}

\item{decimals}{Numero di decimali da mostrare.}
}
\value{
Una stringa rappresentante un numero arrotondato con i decimali indicati.
}
\description{
Funzione per rappresentare un numero arrotondato ai decimali richiesti.
A partire da un numero si ottiene una stringa.
}
