% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.portfolios.R
\name{assign.portfolios}
\alias{assign.portfolios}
\title{Assign Portfolios}
\usage{
assign.portfolios(
  data,
  break_var,
  sort_var,
  bps = NULL,
  n_portfolios = 10,
  P_prefix = "P",
  bp_prefix = "bp"
)
}
\arguments{
\item{data}{A tibble that contains at least two variables, one is used to compute breakpoints and the other is used to compute portfolio membership.}

\item{break_var}{The variable that is used to compute breakpoints.}

\item{sort_var}{The variable that is used to compute portfolio membership.}

\item{bps}{A numerical vector specifying customised preakpoints.
It should contain only the breakpoints not the end values.
For example, for a tercile sort, this value should be c(0.2, 0.8), not c(0, 0.2, 0.8, 1).}

\item{n_portfolios}{A possitive number.
This is the umber of portfolios to which assets are evenly assigned.
Breakpoints are computed evenly across the range in each cross-section.}

\item{P_prefix}{The prefix to be placed in front of portfolio numbers.
This will be part of the column names for the portfolio membership columns.
The default is 'P', which gives column names P1, P2, P3, ...}

\item{bp_prefix}{Similar to P_prefix, this is the prefix to be placed in front of breakpoint numbers.
This will be part of the column names for the breakpoint columns.
The default is 'bp', which gives column names bp1, bp2, ...}
}
\value{
A tibble that contains two types of columns (variables), one is the portfolio membership columns, the other is the breakpoint columns.
Each portfolio membership column gives Boolean values indicating whether an asset belongs to that portfolio.
Each breakpint column gives the breakpoint used for computing membership, this will be one column more than the portfolio memebership columns.
}
\description{
This function assigns each of the assets to a portfolio in the cross-section.
This function implements the portfolio sorting method as in Bali et al. (2016).
That is, it is possible for an asset to be assigned to more than one portfolios in a given cross-section.
}
