% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bland.altman.R
\name{bland.altman}
\alias{bland.altman}
\title{bland.altman}
\usage{
bland.altman(x, y, ylim = c(), pch = c(), main = c(), xlab = c(), ylab = c())
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}

\item{ylim}{Y axis limits}

\item{pch}{pch as in plot}

\item{main}{Title of the plot}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}
}
\value{
Bland altman plot
}
\description{
bland.altman
}
