% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_paramter_bootstrap_function.R
\name{boot_param_CI}
\alias{boot_param_CI}
\title{lme4 and glmmTMB
Poisson
negative binomial
binomial
zero inflated Poisson
zero inflated negative binomial}
\usage{
boot_param_CI(nsim, model, data)
}
\description{
lme4 and glmmTMB
Poisson
negative binomial
binomial
zero inflated Poisson
zero inflated negative binomial
}
