% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conf_inter.R
\name{conf_inter}
\alias{conf_inter}
\title{add confidence intervals to a plot}
\usage{
conf_inter(x,y, stder=FALSE, lowci, upci)
}
\arguments{
\item{x}{vectors of values on the x axis}

\item{y}{vectors of values on the y axis}

\item{stder}{TRUE if you have SE values (1.96*SE), FALSE if you already have CI limits}

\item{low_ci}{vector with lower confidence limits}

\item{up_ci}{vector with upper confidence limits}

\item{z}{how large should the end of the forizontal line of the confidence interval, default is 0.1}

\item{col}{color, default is black}

\item{lty}{line type, default is 1}

\item{lwd}{line width, default is1}

\item{SE}{vectors of values on the for the standard error if stder=TRUE}
}
\description{
Provide x and y vectors of values and if you have Standard error values or upper condidence intervals and lower confidence itnervals
}
\author{
Cyril Milleret
}

