% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{confocal.heatmap}
\alias{confocal.heatmap}
\title{Title: Function to automatically make a heatmap}
\usage{
If replicas present, `confocal.heatmap()` is used best in conjunction with \code{\link{match.pat}} so that
it can determine the replicas within the file vector and add the indices accordingly.
If no replicas, then it can be used as a stand alone.
}
\arguments{
\item{selected_files}{A string vector with the files selected for mapping}

\item{replicas}{A logical value indicating whether the vector has replicas that should be averages.
If T, the function will take the average from the files indicates to by replicas by `index = `.
If F, then no average will be calculated and plots will be generated for each file in the `selected_files` vector.}

\item{index}{A numeric vector e.g. `c(1,2)` indicating the location of the replicas within the `selected_files` vector.}
}
\value{
A plot saved within `02_fig_output`.
}
\description{
\code{confocal.heatmap} It automatically makes a heatmap from a dataframe.
}
\examples{
library(tidyverse)
library(ggprism)
example_data <- data.frame(x = c(1, 2, 3), y = c(3, 67, 90))
get.rsquared(example_data, "x", "y")

}
