% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanMouseConversion.R
\name{convertHumanGeneList}
\alias{convertHumanGeneList}
\title{Basic function to convert human to mouse gene names}
\usage{
convertHumanGeneList(x)
}
\description{
Basic function to convert human to mouse gene names
}
