% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouseHumanConversion.R
\name{convertMouseGeneList}
\alias{convertMouseGeneList}
\title{Basic function to convert mouse to human gene names
This function converts a vector of mouse gene symbols to human gene symbols.}
\usage{
convertMouseGeneList(x)
}
\arguments{
\item{x}{A vector of mouse genes}
}
\description{
Basic function to convert mouse to human gene names
This function converts a vector of mouse gene symbols to human gene symbols.
}
