% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{covariance}
\alias{covariance}
\title{covariance}
\usage{
covariance(x)
}
\arguments{
\item{x}{data frame or matrix with the variables to calculate the covariance}
}
\value{
Matrix with covariance values
}
\description{
Calculates the covariance across a set of covariates for Analysis of Compositional Data
}
