% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRSP.MERGE.R
\name{CRSP.MERGE}
\alias{CRSP.MERGE}
\title{Merging WRDS CRSP stock file and event file}
\usage{
CRSP.MERGE(
  freq = "d",
  vars = c("prc", "ret", "shrout", "ticker", "ncusip", "exchcd", "shrcd", "siccd"),
  START_date = "2022-01-01",
  END_date = "2022-09-27",
  usnm = "username",
  pwd = "password",
  pgpass = FALSE
)
}
\arguments{
\item{freq}{The frequency of the data being retrieved. Passible values are 'd' for daily and 'm' for monthly.}

\item{vars}{A character vector specifying variable names from the stock file.}

\item{START_date}{The start date of the sample, in the format 'YYYY-MM-DD'.}

\item{END_date}{The end date of the sample, in the format 'YYYY-MM-DD'.}

\item{usnm}{WRDS username.}

\item{pwd}{WRDS password. When pgpass is TRUE, this argument can be ignored.}

\item{pgpass}{A boolean value indicating whether a .pgpass file has been set up on local machine.}
}
\value{
A tibble that contains the merged data.
}
\description{
This function merges the stock file and the event file of CRSP.
It is essentially an R translation of the SAS macro CRSPMERGE by WRDS:https://wrds-www.wharton.upenn.edu/pages/support/research-wrds/macros/wrds-macro-crspmerge/, with some minor changes.
I thank the authors of the original SAS code for granting the right of using their logic.
It requires R version 4.2.0 or above as it uses the native pipe and its placeholder.
}
