% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreadspatial.R
\name{dbreadspatial}
\alias{dbreadspatial}
\title{Read a  spatial vector data from postgres}
\usage{
dbwritespatial(con, myspatialdata.spdf, schemaname="myschema", tablename="myspatialdata", replace=FALSE)
}
\arguments{
\item{con}{set the connection with the database:  con <- dbConnect(dbDriver("PostgreSQL"), dbname="postgis", port=5432, host="localhost", user="postgres", password="pw")}

\item{schemaname}{name of the schema in which the table is stored. 'public' is the default}

\item{tablename}{string name of the table to be imported}

\item{geomcol}{the field name in which geometry info are stored}

\item{idcol}{Name of the column with unique IDs to be used in the ID slot of the spatial objects (not relevant for point data)}
}
\description{
it returns the spatialPolygonDataFrame with its projection, but it only works if there is one single projection system in the table
}
\author{
Philipp Hunziker <https://philipphunziker.wordpress.com/2014/07/20/transferring-vector-data-between-postgis-and-r/>
}

