% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbwritespatial.R
\name{dbwritespatial}
\alias{dbwritespatial}
\title{Write spatial vector data in postgres}
\usage{
dbwritespatial(con, myspatialdata.spdf, schemaname="myschema", tablename="myspatialdata", replace=FALSE)
}
\arguments{
\item{con}{set the connection with the database:  con <- dbConnect(dbDriver("PostgreSQL"), dbname="postgis", port=5432, host="localhost", user="postgres", password="pw")}

\item{spatial.df}{spatial dataframe to be sent in postgres}

\item{SRID}{set the SRID number ; e.g 3021 is for RT90}

\item{replace}{if FALSE existing table is not replaced. True table is replaced}

\item{name}{name of the new table to be created}
}
\description{
Write spatial vector data in postgres
}
\author{
Philipp Hunziker <https://philipphunziker.wordpress.com/2014/07/20/transferring-vector-data-between-postgis-and-r/>
}

