% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonal.time.series.OLS.R
\name{diagonal.time.series.OLS}
\alias{diagonal.time.series.OLS}
\title{Generating a diagonal-style table from time-series OLS analysis}
\usage{
diagonal.time.series.OLS(Dep, Indep, lagged = FALSE, ...)
}
\arguments{
\item{Dep}{A xts which is the dependent variable.}

\item{Indep}{A named list of xts objects which are independent variables.}

\item{lagged}{A Boolean value indicating whether to lag the independent variable by one lag.}

\item{...}{Other arguments taken by OLS.REG.}
}
\description{
This function produces a table where indipendent variables are used by the OLS analysis first once a time and then all together.
}
