% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_app.R
\name{DNA_app}
\alias{DNA_app}
\title{Open an interactive map of DNA samples collected in your browser}
\usage{
DNA_app(WD)
}
\arguments{
\item{WD}{character string; folder location of the files "DNA_sample.csv" and "MCP_98_12.shp" e.g C:/Users/cyrilm/Desktop"}
}
\value{
interactive map of DNA samples of the individual selected
}
\description{
Open an interactive map of DNA samples collected in your browser
}
\examples{
DNA_app("C:/Users/cyrilm/Desktop") 
}

