% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DNA_plot.R
\name{DNA_plot}
\alias{DNA_plot}
\title{Open an interactive map of DNA samples collected in your browser}
\usage{
DNA_plot(ID="G45-12",winter="1112",WD="C:/My_documents/Phd_Cyril/DNA/interactive_map/function_plot")
}
\arguments{
\item{ID}{character string; ID of the Individual}

\item{winter}{character string; winter of the MCP e.g. for MCP 2007/2008 write "0708"; for 2000/2001, write "0001"}

\item{WD}{character string; folder location of the files DNA_sample.csv and MCP_98_12.shp}
}
\value{
interactive map of DNA samples of the individual selected
}
\description{
Open an interactive map of DNA samples collected in your browser
}

