% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseliners.R
\name{elastic_baseline}
\alias{elastic_baseline}
\title{Signal baseline}
\usage{
elastic_baseline(x, n_points = 20L, k = 5L, filter_lows = TRUE)
}
\arguments{
\item{x}{`numeric`: time series to find the baseline.}

\item{n_points}{`integer`: number of points to calculate the baseline.}

\item{k}{`integer`: number of points to clear each iteration.}

\item{filter_lows}{`logic`: whether or do a linear fitting to remove values
above the trend.}
}
\description{
Signal baseline
}
