% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{elvt_auc}
\alias{elvt_auc}
\title{AUC}
\usage{
elvt_auc(actual, prediction, segment = NULL)
}
\arguments{
\item{actual}{List of actual values - 0s and 1s of the target variable}

\item{prediction}{List of predicted probabilities}

\item{segment}{Optional parameter. One segment variable can be passed if the auc has to be calculated within each segment}
}
\value{
If no segment is passed, returns a numberic value that is AUC   \cr  
If segment is passed, a data frame containing segment and correponding AUC is returned. One row of overall AUC is also returned
}
\description{
This function calculates the AUC score for the given actual and predictions. If auc needs to be calculated within each segment such as state or vintage, passing the segment to the parameter segment would return AUC by each segment. At present only one segment variable can be used at a time.
}

