% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{elvt_auc_drop_rate}
\alias{elvt_auc_drop_rate}
\title{AUC drop rate}
\usage{
elvt_auc_drop_rate(train_auc, test_auc)
}
\arguments{
\item{train_auc}{Numeric value of AUC from train data. This could CV train or overall training data auc}

\item{test_auc}{Numeric value of AUC from test data. This could CV test or validation auc}
}
\value{
Percentage of drop seen in AUC
}
\description{
Calculates the AUC drop rate between train and test AUCs. Ideally we would want to keep this value less than 10% in order to avoid over fit.
}

