% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{elvt_cap_floor}
\alias{elvt_cap_floor}
\title{Capping and flooring}
\usage{
elvt_cap_floor(x, train_ind = NULL, floor = T)
}
\arguments{
\item{x}{List of numerical values which needs to be capped and floored}

\item{train_ind}{Logical list with length equal to x, indicating the values of x that belong to train. When provided, transformations are done across samples - i.e training data to identify interquartile range and values to replace and apply them on both train and test. Defaults to NULL which will use the entire distribution to perform transformation.}

\item{floor}{Logical indicator for should flooring also be done. Defaults to TRUE}
}
\value{
A list with length same as x containing transformed values
}
\description{
Does capping and if required flooring one variable at a time. Any values outside 1.5 times inter quartile range is capped and floored with values at 95th and 5th percentile respectively. Can also be done across folds (i.e use training data to identify interquartile range and values to replace and apply them on both train and test). If standard deviation becomes 0 after cap and floor, which can happen with variables that have very few unique values, the original variable is returned without transformation.
}

