% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{elvt_char_numeric_in_same_col}
\alias{elvt_char_numeric_in_same_col}
\title{Checks if character and numeric values are in same column}
\usage{
elvt_char_numeric_in_same_col(df)
}
\arguments{
\item{df}{Data frame that needs to be checked}
}
\value{
list of names of columns that has both character and numeric values in it.
}
\description{
This function checks if any of the columns in a data frame has both number and character values. There is potential to capture parsing issue this way, but could also raise lot of false flags.
}

