% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{elvt_col_with_delimiter}
\alias{elvt_col_with_delimiter}
\title{Delimiter present in the data?}
\usage{
elvt_col_with_delimiter(df)
}
\arguments{
\item{df}{R data frame that needs to be checked for presence of delimitor}
}
\value{
List of column names (if any) that have common delimitors.
}
\description{
When writing data frame to delimited flat file, the data should be checked for commonly used delimitor presence in the data. For instance if a column has user entered address which might have "," in them, then comma should be avoided as a delimitor to write the file. This function takes a data frame and returns list of variable names if any that have commonly used delimitors in them.
}

