% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dev.R
\name{elvt_compress_tar}
\alias{elvt_compress_tar}
\title{Create tar file}
\usage{
elvt_compress_tar(dirName)
}
\arguments{
\item{dirName}{Complete path of the directory with all the package files ready to be compressed}
}
\value{
Comresses folder to tar.gz file
}
\description{
Takes all the package files and creates tar file from it
}

