% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{elvt_decimal_to_percent}
\alias{elvt_decimal_to_percent}
\title{Decimal to percent}
\usage{
elvt_decimal_to_percent(x, digits = 2)
}
\arguments{
\item{x}{integer or list of interger that need to be formated as percentage}

\item{digits}{Number of digits after decimal point to retain. Defaults to 2}
}
\value{
The list of decimal values formatted as percentage string.
}
\description{
Function takes a decimal value and returns it formatted as percentage.
}

