% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{elvt_delimiter_inside_field}
\alias{elvt_delimiter_inside_field}
\title{Check if delimiter is inside any of the fields}
\usage{
elvt_delimiter_inside_field(fileName, fieldSeperator, nrows = NULL)
}
\arguments{
\item{fileName}{Complete path of the flat file to be analysed}

\item{fieldSeperator}{The field seperator used in the file}

\item{nrows}{First n rows to be looked into. In case of large files using this parameter could speed things up by looking at only sample of rows. Defaults to NULL which looks at all the rows}
}
\value{
Prints whether the field seperator is present in any of the fields along number of maximum and minimum number of columns identified.
}
\description{
This function takes flat file as input and checks if the delimitor is present anywhere inside any of the fields. 
This will help prevent parsing issue in cases such as when there are pipe seperated values present within a column in a pipe delimited file. 
   Methodology - The function calculates number of columns at each row and returns the max & min values across the entire file. Any difference between max and min indicates issue.
}

