% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{elvt_down_sample_stratified}
\alias{elvt_down_sample_stratified}
\title{Sub sampling with fixed target}
\usage{
elvt_down_sample_stratified(df, target_ratio, targetCol, strata)
}
\arguments{
\item{df}{data framet that needs to be sampled}

\item{target_ratio}{Integer 0-1 of the expected target ratio in the sampled data. This should be more than current target ratio or else all records will be included for current strata and target ratio can not be maintained}

\item{targetCol}{Target column (response, fpd, chargeoff etc..)}

\item{strata}{list of column names by which the data needs to be stratified}
}
\value{
List of row numbers that needs to be retained in the current data set
}
\description{
Down samples the majority class (assumed to take value 0; with minor = 1) while retaining all minor samples, to the target_ratio provided, stratified by multiple columns. \cr
The target_ratio provided should be higher than actual ratio in the unsampled data. If this is not true, no sampling for that strata is done and expected target ratio can not be maintained.
}

