% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{elvt_getSummary}
\alias{elvt_getSummary}
\title{Data summary of data frame}
\usage{
elvt_getSummary(dat, vintage = NULL)
}
\arguments{
\item{dat}{Data frame for which summary statistics need to be calculated}

\item{vintage}{Optional parameter to provide name of the variable in dat that indicates vintage. If this is provided then anova test is also done}
}
\value{
Returns a dataframe with one row for each column contatining following summary statistics. Some of these values might not be present depending on the data type of the variable under consideration.            \cr
         1. col_name - Name of the column \cr    
         2. data_type - Data type deducted by R \cr   
         3. Percent_missing - Percentage of rows that are missing  \cr  
         4. Unique_values - Number of unique values \cr    
         5. Percent_non0_nonNA - Percentage of values that are non null or 0  \cr   
         6. Mean - Mean of the column   \cr 
         7. SD - Standard deviation \cr    
         8. Min - Minumum value   \cr  
         9. Max - Maximum value \cr    
         10. Anova_p - P value of the anova test done with vintage as target (if vintage parameter is set) \cr
}
\description{
This function takes data frame and returns multiple summary statistics. If vintage parameter is provided it also returns p value of anova test done with vintage variable as target variable. This value is very helpful in deducting month on month stability of the variable. A value of 0 indicates high fluctuation.
}

