% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{elvt_handleCharBinaries}
\alias{elvt_handleCharBinaries}
\title{Handle binary variables present as character}
\usage{
elvt_handleCharBinaries(df, colsToLeave = NULL)
}
\arguments{
\item{df}{The working data frame, character variables would be filtered inside the function}

\item{colsToLeave}{Optional parameter to provide list of variables that needs to be left alone}
}
\value{
Original data frame with all binary characters replaced by 1/0s
}
\description{
Bureau variables on occasions has binary variables encoded as true/false, y/n, green,blue. This function turns all those variables into 1s & 0s. \cr  
colsToLeave parameter can be used to leave any variables that you might want to not replace
}

