% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dev.R
\name{elvt_install_elvtFunctions}
\alias{elvt_install_elvtFunctions}
\title{Install elvtFunctions on all R versions}
\usage{
elvt_install_elvtFunctions(package_location, tar = F)
}
\arguments{
\item{package_location}{Either tar.gz file to install from or complete package folder (in which case tar should be set as T)}

\item{tar}{If TRUE then the package_location is a directlory name and is compressed into tar.gz before installing}
}
\description{
Installs package on all R versions (currently 3.3.0, 3.3.1, 3.3.2 are included)
If the package is not compressed into tar.gz, setting parameter tar = T will compress too
}

