% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_keeping.R
\name{elvt_list_files}
\alias{elvt_list_files}
\title{List files and utilization}
\usage{
elvt_list_files(dir = "~")
}
\arguments{
\item{dir}{Directory under which to look. The function looks at all the sub directories and files recursively. Defaults to home directory.}
}
\value{
Returns a dataframe with following columns  \cr 
1. fileName - Name of the file/directory; \cr           
2. directory - The file path of the above file;\cr          
3. isDirectory - Logical indicator of whether the "fileName" is a directory. TRUE for directory and FALSE if it is a file;          \cr
4. access_date - The date it was last accessed;  \cr        
5. access_time - The timestamp of last access time;  \cr        
6. fileSize - File size in bytes - which is easier for sorting; \cr          
7. fileSize_formatted - File size in human readable form;  \cr
}
\description{
This function returns a dataframe containing list of all the files and directories found under the given directory (recursive), with file size and last access time.
}

