% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{elvt_match_names}
\alias{elvt_match_names}
\title{Match strings that vary slightly}
\usage{
elvt_match_names(nm1, nm2)
}
\arguments{
\item{nm1}{List of names that needs to be matched with nm2. Ensure that nm1 is generally longer in length than nm2.}

\item{nm2}{List of names that needs to be matched with nm1}
}
\value{
Data frame with one row for each names in first parameter nm1 and 5 columns with the best 5 matches for each of them.
}
\description{
Often the same bureau variables have different names depending on the data source. This function matches these subtly different names using restricted Damerau-Levenshtein string distnace and returns top 5 matches for each name. \cr
Note: Always ensure that generally longer version of string names are in nm1. Match rate turns bad if this condition is not met.
}

