% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{elvt_max_ks_score}
\alias{elvt_max_ks_score}
\title{Max KS score}
\usage{
elvt_max_ks_score(actual, prediction, top_N_percent = 1)
}
\arguments{
\item{actual}{List of actual values - 0s and 1s of the target variable}

\item{prediction}{List of predicted probabilities}

\item{top_N_percent}{Optional parameter. Should have values between 0 and 1. When provided, returns the max ks within the top n percent of the predicted probability.}
}
\value{
Max KS score.
}
\description{
Function returns the maximum KS score. If top_N_percent parameter is supplied then returns the max KS within the top n percentile of the predictions.
}

