% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{elvt_one_hot_encode}
\alias{elvt_one_hot_encode}
\title{One hot encoding (dummy var creation)}
\usage{
elvt_one_hot_encode(df, level_cutoff = 1000, colsToLeave = NULL)
}
\arguments{
\item{df}{The master data frame. It can also contain variables of other types which would be ignored}

\item{level_cutoff}{Any variable that has more than this number of levels are left untouched. This is to avoid the data set size blowing up. Defaults to 1000}

\item{colsToLeave}{List of column names that needs to left alone. Defaults to NULL}
}
\value{
Returns data frame with OHEs created for character variables and the original variables removed
}
\description{
Takes a dataframe and creates one hot encoding (dummy variables) for all character variables in it. Function creates one variable for each level of character variable. The new names will be of the patter "OriginalVariableName_level". Missing values get their own one hot encodes with "_missing" appeneded to variable name.
}

