% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{elvt_plot_logit_curve}
\alias{elvt_plot_logit_curve}
\title{Plot logit curve}
\usage{
elvt_plot_logit_curve(df, varName, target)
}
\arguments{
\item{df}{Dataframe containing the target as well as variable to plot}

\item{varName}{Name of the variable that has to be plotted}

\item{target}{Name of the target variable}
}
\value{
Logit plot as output
}
\description{
Plot the logit curve along with dot histogram of the actual target values.
}

