% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{elvt_replace_all_NAs}
\alias{elvt_replace_all_NAs}
\title{Replace all NAs in data frame}
\usage{
elvt_replace_all_NAs(df, replace_with = -999)
}
\arguments{
\item{df}{Dataframe with all the columns for which missing values need to be replaced}

\item{replace_with}{Value that the missing values should be replaced with. Defaults to -999}
}
\value{
Returns the original data frame with all NAs replaced with the provided value
}
\description{
Replaces all missing values in the data frame with provided value - this df[is.na(df)] <- -999 fails where there are too many rows
}

