% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{elvt_replace_empty_string_with_NA}
\alias{elvt_replace_empty_string_with_NA}
\title{Replace empty string with NA}
\usage{
elvt_replace_empty_string_with_NA(df)
}
\arguments{
\item{df}{Data frame with all the columns for which missing values need to be replaced with Nas}
}
\value{
The data frame passed to the function with all empty strings replaced by NAs
}
\description{
Replaces the empty string values in the data frame with missing values.
}

