% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{elvt_rmv_tbl_name_prefix}
\alias{elvt_rmv_tbl_name_prefix}
\title{Remove table name prefix}
\usage{
elvt_rmv_tbl_name_prefix(x)
}
\arguments{
\item{x}{Data frame which needs table name prefix to be removed}
}
\value{
The original data frame with prefix stripped off from the names
}
\description{
Often when pulling data from hive, the table name gets prefixed to the column names. This function removes such prefix if any (if there are no prefix, this function doesn't affect the column names)
}

