% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_keeping.R
\name{elvt_search_R_code}
\alias{elvt_search_R_code}
\title{Search inside R codes}
\usage{
elvt_search_R_code(dir = "~", search_phrase, function_declaration = F)
}
\arguments{
\item{dir}{Directory under which to look. The function looks at all the sub directories and files recursively. Defaults to home directory.}

\item{search_phrase}{Search phrase that needs to looked in R files. Regex can also be used.}

\item{function_declaration}{Logical value. If set to T then search_phrase is assumed to be function name and the file where it is declared is returned}
}
\value{
A list of all the files which has the given search term
}
\description{
This function searches in all the R files under the provided directory and returns list of files that has the search term required. 
This is particularly usefull when looking for functions used in past but not sure where the file is or to see if a particular function has been used by anyone else and present under esas folder. 
This can search any .R* files.
}

