% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{elvt_stratified_sample}
\alias{elvt_stratified_sample}
\title{Stratified sampling}
\usage{
elvt_stratified_sample(strata, split_ratio, seed = 1)
}
\arguments{
\item{strata}{Data frame containing variables or a single variable that should be used for stratification. Include the target/response/model dependent variable to maintain constant target rate across samples}

\item{split_ratio}{A list with length equaling number of samples required, containing proportions of rows needed in each sample. If adds to < 1 then another sample for the balance value is added. If adds > 1 then only they are normalized to 1}

\item{seed}{Seed to set. Default is 1}
}
\value{
A list of numbers with length equal to number of rows of data provided. nth value in the list corresponds to the sample nth row of the data belongs to.
}
\description{
Splits the data into any number of samples, holding the distribution of any number of variables (passed as parameter "strata") constant across splits. Pass the target variable to maintain the target rate same across all the splits
}

