% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{elvt_univariate_best_split}
\alias{elvt_univariate_best_split}
\title{Univariate split for rules}
\usage{
elvt_univariate_best_split(df, target, min_percent_rows_needed,
  min_target_rate_needed, selFeats = NULL)
}
\arguments{
\item{df}{Data frame with all the variables and target needed for splitting}

\item{target}{Name of the target variable}

\item{min_percent_rows_needed}{What is the minumum percentage of rows needed to make a split. This is after removing the missing rows}

\item{min_target_rate_needed}{Minimum acceptable target rate to be present in the split to be listed in the final result}
}
\value{
A dataframe with following columns for splits that meets the constraints         \cr
1. feature - Name of the feature    \cr
2. value - Value at which to make the split   \cr 
3. percent_cvrd - Percentage of rows that can be rejected due to this split    \cr
4. bad_rate - Bad rate of the rejected population by this split \cr
}
\description{
This function is used to identify univariate rules that provides maximum seperation between goods and bads in the training data. The function returns list of variables, the best cut points in them that satisfies the constraints. The 2 constraints are - Minimum % of rows needed and minimum % of target needed - which ensures that there are certain minimum percentage of rows present when the data is split and also a certain threshold for target rate. Register a backend to automatically perform operations in parallel.
}

