% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{elvt_unsupervised_feat_reduction}
\alias{elvt_unsupervised_feat_reduction}
\title{Unsupervised feature reduction}
\usage{
elvt_unsupervised_feat_reduction(df, excludeCols = NULL,
  nonMissingRate = NULL, varianceCutOff = NULL, returnDropReasons = F)
}
\arguments{
\item{df}{Master data frame with all columns}

\item{excludeCols}{List of columns to be excluded from feature reduction. This contains any IDs, target variable or other non independent variable for the model}

\item{nonMissingRate}{Remove any variables with \% of non missing rows, non empty or non zero values less than this}

\item{varianceCutOff}{Removes variables that do not have much variance with just one value representing large \% of records. Any variables that has (1 - \% of most frequent variable in non missing rows) less than varianceCutOff is removed}

\item{returnDropReasons}{Logical value. If TRUE a dataframe of variables removed along with reason removed returned along with the reduced data set}
}
\value{
If returnDropReasons is set to FALSE, the reduced data frame after features removed would be returned. All excludeCols are also added to the final data frame returned    \cr
If returnDropReasons is set to TRUE, the reduced data frame and another data frame with list of removed variables with reasons removed is returned. All excludeCols are also added to the final data frame returned
}
\description{
Performs unsupervised feature reduction on the data set based on following criteria   \cr        
1. Variables with 0 variance are removed by default    \cr 
2. Variables with only 0s and NAs are removed by default   \cr  
3. Variables with low percentage of non missing or zero rows if nonMissingRate parameter is set \cr 
4. Variables with low variance (percentage of non most frequence values) if varianceCutOff parameter is set \cr
}

