% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{explore}
\usage{
explore(dat = c(), idvar = c(), outputdir = "./")
}
\arguments{
\item{dat}{Dataset}

\item{idvar}{Subject identification variable}

\item{outputdir}{Directory to store the output}
}
\value{
Nothing, it saves a csv and pdf file with descriptives and visualizations
}
\description{
Calculates descriptives and visualizations of distributions of variables in dataset
}
