% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{filltime}
\alias{filltime}
\title{Function to fill in times at a certain freq}
\usage{
filltime(startt, endt, freq, updown)
}
\arguments{
\item{startt}{Start datetime (Posix)}

\item{endt}{End datetime (Posix)}

\item{freq}{Spacing of times to fill in (must be in seconds!)}

\item{updown}{If updown == "up" times are rounded up to the nearest second; otherwise down}
}
\description{
Useful to create a continuous time vec for interpolation
}
