% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseliners.R
\name{find_baseline}
\alias{find_baseline}
\title{Signal baseline}
\usage{
find_baseline(x, n_points = 20L, k = 5L)
}
\arguments{
\item{x}{numeric: time series to find the baseline.}

\item{n_points}{integer: number of points to calculate the baseline.}

\item{k}{integer: number of points to clear each iteration.}
}
\description{
Function to find the reference baseline in a
fluorescence traceplot. The algorithm operates as follows:
- First, it searches for the minimum value within the time series.
- Second, removes $k$-neightbours before and after.
- Third repeats $n$-times.
- Fourth, fit a linear regression line to the filtered points.
}
