% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileImporters.R
\name{Fl_PlateReader}
\alias{Fl_PlateReader}
\title{Data importer from plate Fluorescent reader in pathology}
\usage{
Fl_PlateReader(file_list = file_list, Rows = Rows, Columns = Columns)
}
\arguments{
\item{file_list}{a list with the file paths in format .txt to be imported}

\item{Rows}{a list containing "n" vectors of lenght 8 with all the conditions across rows in the plate}

\item{Columns}{a list containing "n" vectors of lenght 12 with all the conditions across columns in the plate}
}
\value{
"D" a data.table containing all the reads, times, and conditions
}
\description{
Function to import the reads of a time serie from the plate reader at the Dr Summers Lab.
}
\author{
Santiago Caño-Muñiz
}
