% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsApply_1col.R
\name{fsApply_1col}
\alias{fsApply_1col}
\title{Apply function over columns in a flowSet}
\usage{
fsApply_1col(x, y = "FSC", FUN = median(), ...)
}
\arguments{
\item{x}{flowSet to be used}

\item{y}{string vector indicating the columns}

\item{FUN}{the function to be applied to each element of x}
}
\description{
like many of the apply-style functions in R acts as an iterator for flowSet objects, allowing the
application of a function to either the flowFrame or the data matrix itself. The output can the be reconstructed
as either a flowSet, a list or a matrix depending on options and the type of objects returned..
}
\author{
Santiago Caño-Muñiz
}
