% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_GS.R
\name{Gaussian_GS}
\alias{Gaussian_GS}
\title{Gibbs Sampling from a normal distribution}
\usage{
Gaussian_GS(x, niter = 2000, prior.mu = 0, prior.sd = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{niter}{Number of iterations to extract samples}

\item{prior.mu}{Starting value of mu}

\item{prior.sd}{Starting value of sigma}

\item{burn.in}{Warm up samples}
}
\description{
Manual implementation of basic gibbs sampling from normal distribution
}
\author{
Santiago Caño-Muñiz
}
