% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.palette.R
\name{get.palette}
\alias{get.palette}
\title{Get a named palette}
\usage{
get.palette("palette_name")
Make sure the name is a string.
}
\arguments{
\item{pal_blind}{different colours from colour-blind friendly palette}

\item{greens}{Dark to light greens}

\item{yellows}{Dark to light yellows}

\item{series_two}{Black and red}

\item{GPT}{Palette from chatGPT}

\item{pastel}{Pastel colours}
}
\description{
Just a list of palettes I've used in the thesis.
}
\examples{
library(tidyverse)
library(ggprism)
example_data <- data.frame(x = c(1:8),
                           y = seq(10, 80, by = 10),
                           z = factor(c(1:8)))
ggplot(example_data, aes(x, y, colour = z)) +
  geom_point() +
  scale_color_manual(values = get.palette("pal_blind")) +
  thesis.theme()
get.palette("pal_blind")
}
