% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.rsquared.R
\name{get.rsquared}
\alias{get.rsquared}
\title{Title: Function to get R2, Sy, and slope}
\usage{
x and y are strings with the respective column names.
}
\arguments{
\item{data}{A dataframe with the x and y values to be calculated.}

\item{x}{A string ("name") indicating the name of the column with the x axis value (independent variable)}

\item{y}{A string ("name") indicating the name of the column with the y axis value (dependent variable)}
}
\value{
A list is generated with the following components:
\item{R2 }{R2 of data set}
\item{b }{b-intercept of linear equation}
\item{slope }{Slope of the linear equation generated by lm}
\item{residuals }{Sy: The sd of the residuals}
}
\description{
\code{get.rsquared} uses the linear modelling function \code{\link{lm}} to calculate the R2, the residual of the standard error (Sy),
and slope from a data set. These are all helpful in defining correlation and calculating limits of detection.
The end result is a list.
}
\examples{
example_data <- data.frame(x = c(1, 2, 3), y = c(3, 67, 90))
get.rsquared(example_data, "x", "y")
}
