% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{get_mypalette}
\alias{get_mypalette}
\title{Get my palette.}
\usage{
get_mypalette(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{char: Name of the colour set. Currently: "contrast", "main", "cool", "hot", "mixed" and "grey".}

\item{reverse}{Logical. Should the order of the colors be reversed?}

\item{...}{
  Arguments passed on to \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette}}
  \describe{
    \item{\code{colors}}{colors to interpolate; must be a valid argument to
    \code{\link[grDevices]{col2rgb}()}.}
  }}
}
\description{
Wrapper function to call the colours from my palettes. Original code from Simon Jackson (twitter:
[at]drsimonj, https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2)
}
\author{
Santiago Caño-Muñiz
}
