% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plus_minus_changes.R
\name{get_plus_minus_changes}
\alias{get_plus_minus_changes}
\title{get_plus_minus_changes}
\usage{
get_plus_minus_changes(
  dataf,
  y,
  comps,
  comps_fup = NULL,
  covars = NULL,
  deltas = c(0, 10, 20)/(24 * 60),
  comparisons = c("one-v-one", "one-v-all", "prop-realloc")[1],
  longAnalysis = c("prospective", "change")[1],
  alpha = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{dataf}{Data frame with the data}

\item{y}{String with the name of the outcome in dataf}

\item{comps}{String with the name of the composition variables in dataf}

\item{comps_fup}{String with the name of the follow-up composition variables in dataf (optional)}

\item{covars}{String with the name of the covariates}

\item{deltas}{A vector of time-component changes (as proportions of compositions, i.e., values between -1 and 1)}

\item{comparisons}{Two choices ("one-v-one" or "prop-realloc")}

\item{longAnalysis}{}

\item{alpha}{Optional, level of significance of results (default = 0.05)}

\item{verbose}{Whether to print output on the progress in the console.}
}
\value{
Data frame with predictions
}
\description{
Get predictions from compositional ilr multiple linear regression model
}
