% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions_coda.R
\name{get_predictions_coda}
\alias{get_predictions_coda}
\title{get_predictions_coda}
\usage{
get_predictions_coda(
  data = c(),
  comp_bl = c(),
  comp_fu = c(),
  comp_names = c(),
  outcome_bl = c(),
  outcome_fu = c(),
  covs = c(),
  comparisons = "prop-realloc",
  increase = c(),
  decrease = c(),
  delta = 30,
  model = "change"
)
}
\arguments{
\item{data}{Data frame with composition (baseline and/or follow up), outcome and covariates}

\item{comp_bl}{names of composition variables in your dataset (baseline)}

\item{comp_fu}{names of composition variables in your dataset (follow up, optional)}

\item{comp_names}{names of composition variables for visualization}

\item{outcome_bl}{Name of outcome at baseline in your dataset}

\item{outcome_fu}{Name of outcome at follow up in your dataset (optional)}

\item{covs}{names of your covariates in your dataset}

\item{comparisons}{Tow options: "one-v-one" OR "prop-realloc"}

\item{increase}{name of the behavior to increase (name for visualization)}

\item{decrease}{name of the behavior to decrease (only in "one-v-one")}

\item{delta}{minutes to reallocate for prediction (e.g., +30 min)}

\item{model}{"cross-sectional", "longitudinal", or "change"}
}
\value{
Matrix with predictions
}
\description{
Get predictions from CoDA models
}
