% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertID.R
\name{getMatrixWithSelectedIds}
\alias{getMatrixWithSelectedIds}
\title{Convert the rownames of a dataframe or matrix to a different ID}
\usage{
getMatrixWithSelectedIds(df, type, keys, db)
}
\arguments{
\item{df}{Dataframe of matrix whose rownames are to be converted.}

\item{type}{Format of converted rownames: e.g. "SYMBOL" or "ENTREZID"}

\item{keys}{Present format of rownames: e.g. "SYMBOL" or "ENTREZID"}

\item{db}{Annotation database: e.g. "org.Hs.eg.db" or "org.Mm.eg.db"}
}
\description{
This function can take any of the columns(org.Hs.eg.db) as type and keys as long as the row names are in the format of the keys argument
}
\examples{
M1entrez <- getMatrixWithSelectedIds(M1symb, type="ENTREZID", keys="SYMBOL", db = org.Mm.eg.db)
}
