% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted_descriptions.R
\name{give_binary_description}
\alias{give_binary_description}
\title{give_binary_description}
\usage{
give_binary_description(x = NA, void_string = "-")
}
\arguments{
\item{x}{Un vettore di fattori (rappresentante una variabile binaria), convertito come tale se numerico.}

\item{void_string}{Stringa da usare se il valore non c'è o non è calcolabile.}
}
\value{
Un vettore con i risultati descrittivi.
}
\description{
Funzione rappresentare le percentuali.
}
