% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted_descriptions.R
\name{give_categorical_description}
\alias{give_categorical_description}
\title{give_categorical_description}
\usage{
give_categorical_description(
  x = NA,
  name = "",
  max_missed = 0.1,
  void_string = "-",
  list_marker = "-"
)
}
\arguments{
\item{x}{Un vettore di fattori, convertito come tale se numerico.}

\item{name}{Stringa da usare come nome della variabile (altrimenti è utilizzata l'etichetta, se presente).}

\item{max_missed}{Percentuale massima ammessa di omissioni.}

\item{void_string}{Stringa da usare se il valore non c'è o non è calcolabile.}

\item{list_marker}{Marcatore di punto elenco (per la lista dei livelli).}
}
\value{
Un data.frame con i risultati descrittivi.
}
\description{
Funzione per descrivere una variabile categoriale (con una tabella relativa a tutti i livelli possibili).
}
