% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_basic_tests.R
\name{give_chisquare}
\alias{give_chisquare}
\title{give_chisquare}
\usage{
give_chisquare(x = NA, y = NA, void_string = "-")
}
\arguments{
\item{x}{Vettore di fattori.}

\item{y}{Vettore di fattori.}

\item{void_string}{Stringa da usare se il valore non c'è o non è calcolabile.}
}
\value{
Un vettore con il risultato del test e con il valore di p risultante.
}
\description{
Test del Chi quadrato.
}
