% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted_descriptions.R
\name{give_continuous_crosstable_2group_b}
\alias{give_continuous_crosstable_2group_b}
\title{give_continuous_crosstable_2group_b}
\usage{
give_continuous_crosstable_2group_b(
  y = NA,
  group = NA,
  name_y = "",
  max_missed = 0.1,
  void_string = "-",
  alpha_value = 0.05,
  multiple_alphas = c(0.05, 0.01, 0.001)
)
}
\arguments{
\item{y}{Un vettore di numerico.}

\item{group}{Un vettore di fattori, convertito come tale se numerico, a 2 livelli.}

\item{name_y}{Stringa da usare come nome della variabile (altrimenti è utilizzata l'etichetta, se presente).}

\item{max_missed}{Percentuale massima ammessa di omissioni.}

\item{void_string}{Stringa da usare se il valore non c'è o non è calcolabile.}

\item{alpha_value}{Valore di alpha per la significatività del test.}

\item{multiple_alphas}{Valori di alpha per gli asterischi sulla significatività del test.}
}
\value{
Un data.frame con i risultati descrittivi (una colonna per livello della variabile con i gruppi).
}
\description{
Funzione per descrivere una variabile continua e il suo incrocio con un'altra con 2 gruppi (con una tabella relativa a tutti i livelli possibili).
}
