% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted_descriptions.R
\name{give_continuous_description}
\alias{give_continuous_description}
\title{give_continuous_description}
\usage{
give_continuous_description(x = NA, void_string = "-")
}
\arguments{
\item{x}{Un vettore numerico (rappresentante una variabile continua).}

\item{void_string}{Stringa da usare se il valore non c'è o non è calcolabile.}
}
\value{
Una stringa rappresentante il valore T, con vari abbellimenti.
}
\description{
Funzione rappresentare le percentuali.
}
