% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_basic_tests.R
\name{give_continuous_test_2}
\alias{give_continuous_test_2}
\title{give_continuous_test_2}
\usage{
give_continuous_test_2(y = NA, group = NA, void_string = "-")
}
\arguments{
\item{y}{Vettore numerico.}

\item{group}{Vettore di fattori, a 2 livelli.}

\item{void_string}{Stringa da usare se il valore non c'è o non è calcolabile.}
}
\value{
Un vettore con il risultato del test e con il valore di p risultante.
}
\description{
Test per categorie (Chi quadro o test di Fisher).
}
