% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_basic_tests.R
\name{give_fisher}
\alias{give_fisher}
\title{give_fisher}
\usage{
give_fisher(x = NA, y = NA, void_string = "-")
}
\arguments{
\item{x}{Vettore di fattori.}

\item{y}{Vettore di fattori.}

\item{void_string}{Stringa da usare se il valore non c'è o non è calcolabile.}
}
\value{
Il p risultante.
}
\description{
Test esatto di Fisher.
}
