% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_manipulations.R
\name{give_mean_confidence_intervals}
\alias{give_mean_confidence_intervals}
\title{give_mean_confidence_intervals}
\usage{
give_mean_confidence_intervals(x, confidence = 0.95, distribution = "z")
}
\arguments{
\item{x}{Vettore numerico.}

\item{confidence}{Ampiezza dell'intervallo di confidenza voluto.}

\item{distribution}{Usare una distribuzione z o t (per campioni piccoli o mal distribuiti).}
}
\value{
Un vettore (utile per inserirlo in un data-frame).
}
\description{
Funzione per ottenere l'intervallo di confidenza della media.
}
