% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{give_nice}
\alias{give_nice}
\title{give_nice}
\usage{
give_nice(
  value = NA,
  decimals = 3,
  text = "",
  with_equal_sign = FALSE,
  with_sign = TRUE,
  min_value = -Inf,
  max_value = Inf,
  void_string = "-"
)
}
\arguments{
\item{value}{Valore numerico.}

\item{decimals}{Numero di decimali da mostrare.}

\item{text}{Testo da mostrare prima del valore.}

\item{with_equal_sign}{Segno di uguale (sempre presente se c'è il testo).}

\item{with_sign}{Segno di +/- davanti al numero.}

\item{min_value}{Minimo valore possibile.}

\item{max_value}{Massimo valore possibile.}

\item{void_string}{Stringa da usare se il numero non c'è.}
}
\value{
Una stringa rappresentante un numero, con eventuale nome del valore e vari abbellimenti.
}
\description{
Funzione per rappresentare un numero arrotondato ai decimali richiesti.
A partire da un numero si ottiene una stringa.
}
