% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{give_nice_p}
\alias{give_nice_p}
\title{give_nice_p}
\usage{
give_nice_p(
  value = NA,
  decimals = 3,
  with_p = TRUE,
  with_equal_sign = FALSE,
  with_stars = TRUE,
  multiple_stars = TRUE,
  alpha = 0.05,
  multiple_alphas = c(0.05, 0.01, 0.001),
  give_only_stars = FALSE,
  void_string = "-"
)
}
\arguments{
\item{value}{Valore numerico.}

\item{decimals}{Numero di decimali da mostrare.}

\item{with_p}{Mostra il 'p', prima del numero.}

\item{with_equal_sign}{Segno di uguale (sempre presente se c'è il testo).}

\item{with_stars}{Mostra un asterisco se statisticamente significativo.}

\item{multiple_stars}{Mostra da uno a tre asterischi se statisticamente significativo.}

\item{alpha}{Significatività statistica (0.050)}

\item{multiple_alphas}{Vettore con tre livelli di significatività statistica da asteriscare.}

\item{give_only_stars}{Non restituisce il valore numerico, ma solo gli asterischi.}

\item{void_string}{Stringa da usare se il numero non c'è.}
}
\value{
Una stringa rappresentante il valore di p, con vari abbellimenti.
}
\description{
Funzione per rappresentare i valori di p.
Comprende gli asterischi per la significatività statistica.
}
