% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{give_nice_percent}
\alias{give_nice_percent}
\title{give_nice_percent}
\usage{
give_nice_percent(
  value = NA,
  decimals = 2,
  text = "",
  percent_sign = TRUE,
  min_value = -100,
  max_value = 100,
  with_equal_sign = FALSE,
  with_sign = FALSE,
  void_string = "-"
)
}
\arguments{
\item{value}{Valore numerico.}

\item{decimals}{Numero di decimali da mostrare.}

\item{text}{Testo da mostrare prima del valore.}

\item{percent_sign}{Segno di percentuale da usare.}

\item{min_value}{Punteggio minimo rappresentabile.}

\item{max_value}{Punteggio massimo rappresentabile.}

\item{with_equal_sign}{Segno di uguale (sempre presente se c'è il testo).}

\item{with_sign}{Segno di +/- davanti al numero.}

\item{void_string}{Stringa da usare se il numero non c'è.}
}
\value{
Una stringa rappresentante il valore T, con vari abbellimenti.
}
\description{
Funzione rappresentare le percentuali.
}
