% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{give_nice_r}
\alias{give_nice_r}
\title{give_nice_r}
\usage{
give_nice_r(
  value = NA,
  decimals = 3,
  with_r = TRUE,
  with_equal_sign = FALSE,
  void_string = "-"
)
}
\arguments{
\item{value}{Valore numerico.}

\item{decimals}{Numero di decimali da mostrare.}

\item{with_r}{Mostra il 'r', prima del numero.}

\item{with_equal_sign}{Segno di uguale (sempre presente se c'è il testo).}

\item{void_string}{Stringa da usare se il numero non c'è.}
}
\value{
Una stringa rappresentante il valore di r, con vari abbellimenti.
}
\description{
Funzione rappresentare i valori di r.
}
