% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{give_nice_standard}
\alias{give_nice_standard}
\title{give_nice_standard}
\usage{
give_nice_standard(
  value = NA,
  decimals = 1,
  text = "Standard",
  standard_mean = 100,
  standard_sd = 15,
  min_value = 0,
  max_value = 200,
  with_equal_sign = FALSE,
  with_sign = FALSE,
  void_string = "-"
)
}
\arguments{
\item{value}{Valore numerico.}

\item{decimals}{Numero di decimali da mostrare.}

\item{text}{Testo da mostrare prima del valore.}

\item{standard_mean}{Media del punteggio standard.}

\item{standard_sd}{SD del punteggio standard.}

\item{min_value}{Punteggio minimo rappresentabile.}

\item{max_value}{Punteggio massimo rappresentabile.}

\item{with_equal_sign}{Segno di uguale (sempre presente se c'è il testo).}

\item{with_sign}{Segno di +/- davanti al numero.}

\item{void_string}{Stringa da usare se il numero non c'è.}
}
\value{
Una stringa rappresentante il valore T, con vari abbellimenti.
}
\description{
Funzione per rappresentare i valori Standard (100, 15).
}
