% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{give_nice_z}
\alias{give_nice_z}
\title{give_nice_z}
\usage{
give_nice_z(
  value = NA,
  decimals = 3,
  text = "",
  standard_mean = 0,
  standard_sd = 1,
  min_z = NA,
  max_z = NA,
  with_equal_sign = FALSE,
  with_sign = TRUE,
  void_string = "-"
)
}
\arguments{
\item{value}{Valore numerico.}

\item{decimals}{Numero di decimali da mostrare.}

\item{text}{Testo da mostrare prima del valore.}

\item{standard_mean}{Media del punteggio standard.}

\item{standard_sd}{SD del punteggio standard.}

\item{min_z}{Minimo in SD dalla media.}

\item{max_z}{Massimo in SD dalla media.}

\item{with_equal_sign}{Segno di uguale (sempre presente se c'è il testo).}

\item{with_sign}{Segno di +/- davanti al numero.}

\item{void_string}{Stringa da usare se il numero non c'è.}
}
\value{
Una stringa rappresentante il valore z, con vari abbellimenti.
}
\description{
Funzione per rappresentare i valori z (ma applicabile anche ad altre standardizzazioni).
}
