% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm.R
\name{gm}
\alias{gm}
\title{gm}
\usage{
gm(x, na.rm = T)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Logical. Whether to ignore NA values.}
}
\value{
Geometric mean of the vector
}
\description{
Calculates the geometric mean of a numeric vector
}
