% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_mean.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{`numeric`: vector to calculate the geometric mean.}

\item{na.rm}{`logic`: should NA values be removed.}
}
\description{
Generic function to calculate the geometric mean of a vector. The geometric mean is defined as the nth root of the product of n numbers:
as described: \deqn{\prod_(i\ =\ 1)^n\ofx_i\ =\ \sqrt(n&x_1\bullet x_i\ldots x_n\ )}
}
\author{
Santiago Caño-Muñiz
}
