% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde2d_filter.R
\name{kde2d_filter}
\alias{kde2d_filter}
\title{2 Dimensional K-means filter by densiy}
\usage{
kde2d_filter(x, y, q = 0.05, n = 200, logic = T)
}
\arguments{
\item{x}{x coordinate of data}

\item{y}{y coordinate of data}

\item{q}{quantile filter. The data within that quantile will be gated}

\item{n}{Number of grid points in each direction. Can be scalar or a length-2 integer vector.}

\item{logic}{If true, the returned value is a logic vector with T or F. If false, the function returns the row of the values within the kernel. Default value is True}
}
\description{
Filter values using a density grid on a two-dimensional kernel density gradient
}
\author{
Santiago Caño-Muñiz
}
