% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde2dCpp.R
\name{kde2dCpp}
\alias{kde2dCpp}
\title{Two-dimensional kernel analysis}
\usage{
kde2dCpp(x, y, h, n = 25, lims = c(range(x), range(y)))
}
\arguments{
\item{x}{y coordinate of data a list of ggplot objects}

\item{y}{x coordinate of data}

\item{h}{vector of bandwidths for x and y directions. Defaults to normal reference bandwidth
(see bandwidth.nrd). A scalar value will be taken to apply to both directions.}

\item{n}{Number of grid points in each direction. Can be scalar or a length-2 integer vector.}

\item{lims}{The limits of the rectangle covered by the grid as c(xl, xu, yl, yu).}
}
\description{
Two-dimensional kernel density estimation with an axis-aligned bivariate normal kernel,
 evaluated on a square grid implemented with some Rcpp steps.
}
\author{
Santiago Caño-Muñiz
}
