% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.cal.R
\name{lag.cal}
\alias{lag.cal}
\title{Computing Optimal Lags for Newey-West tests}
\usage{
lag.cal(x)
}
\arguments{
\item{x}{The number of periods in the time series.}
}
\value{
The optimal lags for Newey-West tests.
}
\description{
This function computes the optimal lags for Newey-West tests.
}
