% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbsPres.R
\name{listToAbsPres}
\alias{listToAbsPres}
\title{Convert a list of e.g. genes into a absence/presence data frame with all genes as rownames and each column a list element.}
\usage{
listToAbsPres(x)
}
\arguments{
\item{x}{A list of vectors containing sets of genes or proteins}
}
\description{
A one indicates the presence of a gene in the particular list element, a zero indicates its absence. Potential use of this function is to attach a list of signatures to an expression matrix for downstream analysis or plotting
}
