% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_coda_long.R
\name{lm_coda_long}
\alias{lm_coda_long}
\title{lm_coda_long}
\usage{
lm_coda_long(
  compo_baseline,
  compo_followup = NULL,
  outcome_baseline,
  outcome_followup = NULL,
  moderator = NULL,
  longAnalysis = c("prospective", "change")[1],
  covariates = NULL,
  scale = FALSE,
  balance = NA,
  data
)
}
\arguments{
\item{compo_baseline}{Variable names for the baseline composition}

\item{compo_followup}{Variable names for the follow-up composition (optional)}

\item{outcome_baseline}{Variable name for the baseline outcome}

\item{outcome_followup}{Variable name for the follow-up outcome (optional)}

\item{moderator}{Moderator variable, if any (optional)}

\item{longAnalysis}{Longitudinal analysis to perform, either prospective or changes for now.}

\item{covariates}{Numeric, matrix, or data frame with the covariates to adjust the models for}

\item{scale}{Logical, whether to standardize or not the output (default = FALSE)}

\item{balance}{Balance between behaviors to be used in the ILR coordinates}

\item{data}{Data frame with data}
}
\value{
List with models fit for each one of the ILR coordinates as leading coordinate
}
\description{
Run compositional linear regression models using isometric log ratios. Option to perform longitudinal models and moderation analyses implemented.
}
