% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_eq.R
\name{lm_eq}
\alias{lm_eq}
\title{Linear model equation}
\usage{
lm_eq(x, y, m = NULL)
}
\arguments{
\item{x}{Independent variable}

\item{y}{Dependent variable}

\item{m}{Independent variable}
}
\value{
a character expression in the form "y = b*x + a"
}
\description{
Returns the equation of a linear model as "y = b*x + a". The function also accepts lm objects.
}
\author{
Santiago Caño-Muñiz
}
