% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_ILR.R
\name{LR_ILR}
\alias{LR_ILR}
\title{LR_ILR}
\usage{
LR_ILR(compo)
}
\arguments{
\item{compo}{Matrix or data frame with a composition}
}
\value{
Matrix with isometric log ratios calculated
}
\description{
Calculates the isometric log ratio coordinates of a composition for Analysis of Compositional Data.
}
