% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_funcs.r
\name{lsos}
\alias{lsos}
\title{lsos}
\usage{
lsos(n = 10, pos = 1, order.by = "Size", decreasing = TRUE)
}
\arguments{
\item{n}{Number of objects to list.}

\item{pos}{Environment position \code{1} is the current (global) environment}

\item{order.by}{Attribute to order by, either \code{size} or \code{class},}

\item{decreasing}{logical.  decreasing or increasing order}
}
\value{
a \code{data.frame}
}
\description{
\code{lsos} summarizes objects in environments by their size
}
\examples{
 lsos()

}
