% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confocal_functions.R
\name{match.pat}
\alias{match.pat}
\title{Title: Find matching file names}
\usage{
match.pat(pattern_list)
}
\arguments{
\item{patterns}{A string vector}
}
\value{
A list is generated with the following components:
\item{indices }{the position of the matching elements within the vector}
\item{patterns }{the matching elements}
}
\description{
\code{match.pat} A function customized for my file system where I have 2 replicas for each sample.
The file names for this will only differ by the replica number e.g. A1, A2, B1, B2. But, sometimes more
complicated than that. So, this function finds the matching samples and groups it by replica numbers.
}
\examples{
example_data <-c("C3T2_spec", "C1T1_spec", "C1T2_spec", "C3T1_spec")
match.pat(example_data)
}
