% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_mclapply.R
\name{Mclapply_Hack}
\alias{Mclapply_Hack}
\title{Allows mclapply function to work in PC
This function will hack the mclappy function to work in PC and use multipl
e cores for the analysis. It was created by someone else and the original version is here: 
https://www.r-bloggers.com/implementing-mclapply-on-windows-a-primer-on-embarrassingly-parallel-computation-on-multicore-systems-with-r/}
\usage{
Mclapply_Hack(...)
}
\arguments{
\item{...}{Nothing, just run the function and the mclapply function will change}
}
\value{
A modified, or not, version of mclapply
}
\description{
Allows mclapply function to work in PC
This function will hack the mclappy function to work in PC and use multipl
e cores for the analysis. It was created by someone else and the original version is here: 
https://www.r-bloggers.com/implementing-mclapply-on-windows-a-primer-on-embarrassingly-parallel-computation-on-multicore-systems-with-r/
}
