% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergespdf.R
\name{mergespdf}
\alias{mergespdf}
\title{Merge two SpatialPointsDataFrame}
\usage{
mergespdf(sp1,sp2)
}
\arguments{
\item{sp1}{first spatialPointsdataFrame}

\item{sp2}{second spatialPointsdataFrame}
}
\value{
a merged SpatialPointsDataFrame. it uses the projection of sp1 if the projection is not null
}
\description{
Merge two SpatialPointsDataFrame
}
\author{
Cyril Milleret
}

