% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msc.R
\name{msc}
\alias{msc}
\title{Title: Multi-element string vector}
\usage{
msc(...)
}
\arguments{
\item{...}{elements to be turned into character type.}
}
\value{
A string vector.
}
\description{
\code{msc} makes each element within a vector into a separate string.
}
\examples{
msc(element1, element2, element3)

# for combining numbers and letters
paste0(msc(44,45), "nm")

}
