% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{multiplot}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)}

\item{plotlist}{ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)}

\item{cols}{numeric scalar giving the number of columns in the layout}

\item{layout}{Matrix with layout for plots. If present, cols is ignored.}
}
\description{
Place multiple \code{ggplot} plots on one page.
}
