% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_chng.R
\name{my_chng}
\alias{my_chng}
\title{Estimates percentage change}
\usage{
my_chng(v1, v2, limit = NA)
}
\arguments{
\item{v1}{Type of path needed "D" for Data, "R", for Result and "G" for generic. You need those paths to be load in the global environment}

\item{v2}{I normally call my data folder the same as the project so it automatically sets to the project name. Alternatives available.}
}
\value{
A numeric value in percentage change from v2 relative to v1
}
\description{
This function simply estimnates a percentage change
of a variable with a limit on the scale
}
