% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_data.R
\name{my_data}
\alias{my_data}
\title{Load and installs packages needed}
\usage{
my_data(data, system = "NA")
}
\arguments{
\item{data}{The name of the dataset you wanna call}

\item{system}{If you want to override the automatic system identification}
}
\value{
Paths to save and load data within a project
}
\description{
This function loads all necessary packages. If the
package is not installed, it will install it first
and the it will load the recently installed package
}
