% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ggtheme_m.R
\name{my_ggtheme_m}
\alias{my_ggtheme_m}
\title{My personal map theme}
\usage{
my_ggtheme_m(leg_pos = "top")
}
\value{
a ggplot theme
}
\description{
This function creates my own ggplot theme to 
standarize multiple map in one document
}
