% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_land_map.R
\name{my_land_map}
\alias{my_land_map}
\title{My personal plot theme}
\usage{
my_land_map(scale = "small")
}
\value{
a ggplot theme
}
\description{
This function creates my own ggplot theme to 
standarize multiple plots in one document
}
