% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_mpa_cedar.R
\name{my_mpa_cedar}
\alias{my_mpa_cedar}
\title{Creates DBEM grid for Compute Canada}
\usage{
my_mpa_cedar(data, column_name, path, scen_name)
}
\arguments{
\item{data}{the dataset containing the grid. Note it must be 259200 observations to match the DBEM grid}

\item{column_name}{the column that is to be transformed to the DBEM}

\item{path}{path were data is to be saved}

\item{scen_name}{data set name, no need to inlcude .txt}
}
\description{
This function transforms an ocean grid to the format readable
by the DBEM in Compute Canada
}
