% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_project.R
\name{my_project}
\alias{my_project}
\title{Creates a project basic structure}
\usage{
my_project(Repo = "HALL2000")
}
\arguments{
\item{Repo}{Main data repository. Currently set to HALL2000}
}
\value{
Creates a series of directories that follow the basic structure of all my projects
}
\description{
This function creates all the directories that I normally
use in all my projects. It creats them im bothe the Github 
R project as rell as the data Repository
}
