% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_sf.R
\name{my_sf}
\alias{my_sf}
\title{Loads a specific shapefile}
\usage{
my_sf(sf_name, crs = "NA", simple = "NA")
}
\arguments{
\item{sf_name}{Main data repository. Currently set to HALL2000}
}
\value{
the requested shapefile
}
\description{
This function simply loads a shapefile. I just created it because
I have frquently uised sf and is annoying to be coping + pasting
every single time
}
