\name{myFunctions-package}
\alias{myFunctions-package}
\alias{myFunctions}
\docType{package}
\title{
A set of commonly used R and C++ functions
}
\description{
A set of commonly used R and C++ functions
}
\details{
\tabular{ll}{
Package: \tab myFunctions\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-24\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
John Tipton

Maintainer: John Tipton <jtipton25@gmail.com>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ Examples ~~
}
