% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MySQL.SELECT.WHERE.R
\name{MySQL.SELECT.WHERE}
\alias{MySQL.SELECT.WHERE}
\title{Selecting from MySQL}
\usage{
MySQL.SELECT.WHERE(what, from.table, myConnection, ..., UNION_BY_FROM = FALSE)
}
\arguments{
\item{from.table}{the MySQL table from which the data is selected}

\item{myConnection}{The MySQL connection}

\item{...}{The part of query that follows 'WHERE', and 'AND'}

\item{UNION_BY_FROM}{A logical argument, indicating whether to use 'UNION ALL'. Check the code for details.}

\item{waht}{What to be selected, part of MySQL query that follows 'SELECT'}
}
\value{
The data fetched from DB.
}
\description{
This function executes the MySQL query SELECT...WHERE...AND...
}
