% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nboot.R
\name{nboot}
\alias{nboot}
\title{Bootstrap from a normal distribution}
\usage{
nboot(x, R)
}
\arguments{
\item{x}{Numeric vector}

\item{R}{Number of samples}
}
\description{
Manual implementation of bootstrap sampling from normal distribution
}
\author{
Santiago Caño-Muñiz
}
