% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_overlap_index.R
\name{niche_overlap_index}
\alias{niche_overlap_index}
\title{Index of habitat niche overalp}
\usage{
niche_overlap_index(X, bin=0.10)
}
\arguments{
\item{X}{a kselect (adehabitatHS) object}

\item{bin}{length of the intervals to calculate proportion of used scores}
}
\value{
A list of paired matrix with a Pianka index calculated on each axis and a Weighted (based on eigen values) Pianka index calculated on all axes
}
\description{
Derive a Pianka index from a kselect object to estimate a degree of habitat niche overlap between species/indviduals
}
\examples{
library(adehabitatHS)
data(puechabonsp)

locs <- puechabonsp$relocs
map <- puechabonsp$map

## compute the home range of animals (e.g. using the minimum convex
## polygon)
pc <- mcp(locs[,"Name"])
hr <- hr.rast(pc, map)

## Compute the number of relocation in each pixel of the map
cp <- count.points(locs[,"Name"], map)

## prepares the data for the kselect analysis
x <- prepksel(map, hr, cp)
tab <- x$t

## We call a new graphic window
## A K-select analysis
acp <- dudi.pca(tab, scannf = FALSE, nf = 4)
kn <- kselect(acp, x$factor, x$weight,
              scannf = FALSE, nf = 4)

# compute habitat niche overlap 
niche_overlap_index(kn,bin=0.1)
}
\author{
Cyril Milleret
}

