% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS.REG.R
\name{OLS.REG}
\alias{OLS.REG}
\title{Simple OLS Regressions}
\usage{
OLS.REG(
  multicolumn_data,
  annualisation_factor = 252,
  rounding_decimal = 2,
  scalar = 1,
  star_paren = TRUE,
  NeweyWest = FALSE
)
}
\arguments{
\item{multicolumn_data}{The data that goes into the regression, with the first column being the dependent variable.}

\item{annualisation_factor}{The multiplier used to annualise the data.
Effectively this value is multiplied with the alpha from the regression using original data.
The default value is 252.}

\item{rounding_decimal}{An positive integer indicating how many decimal to keep in the reported results.}

\item{scalar}{A scalar applied to the betas. The default value is 1.}

\item{star_paren}{A Boolean value, indicating whether to give stars to the coefficients.
Note, if TRUE, this changes the results from numeric to character.}
}
\value{
A dataframe that reports the regression results.
}
\description{
This function runs simple OLS regressions and return the results
}
