% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{outliers}
\usage{
outliers(x, k = 1.5)
}
\arguments{
\item{x}{Numeric vector}

\item{k}{Constant to consider outliers (default = 1.5)}
}
\value{
Number of outliers in variable
}
\description{
Identify outliers in a given numeric vector
}
