% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_curves.R
\name{overlap_curves}
\alias{overlap_curves}
\title{Calculate area of overalp between two probability density function}
\usage{
overlap_curves(da,db)
}
\arguments{
\item{da}{a density object See \code{density()}}

\item{db}{a density object See \code{density()}}
}
\value{
the area of overlap (percentages)   between the 2 curves
}
\description{
Calculate area of overalp between two probability density function
}
\author{
Cyril Milleret
}

