% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_inrange.R
\name{peak_inrange}
\alias{peak_inrange}
\title{Peak In range}
\usage{
peak_inrange(x, n1 = 100L, n2 = 10L, peak_loc)
}
\arguments{
\item{x}{numeric: time series to find the baseline.}

\item{n1}{integer: number of points for the slow average.}

\item{n2}{integer: number of points for teh fast average.}

\item{peak_loc}{integer vector: expected location of the peaks.}
}
\value{
A data frame with the peaks and peak ID
}
\description{
Strategy to define peaks within an expected region. The algorithm
uses the ration between a fast average and a slow average to define areas avobe
the normal baseline. Those areas are identified as peaks and then labelled sequentially.
}
