% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileImporters.R
\name{Plate_readerStatic}
\alias{Plate_readerStatic}
\title{Data importer from plate reader}
\usage{
Plate_readerStatic(
  file_list = file_list,
  wl = wl,
  tp = tp,
  Rows = Rows,
  Columns = Columns
)
}
\arguments{
\item{file_list}{a list with the file paths in format .txt to be imported}

\item{wl}{vector with all the wavelenght measured}

\item{tp}{an integer with the total number of time points}

\item{Rows}{a list containing "n" vectors of lenght 8 with all the conditions across rows in the plate}

\item{Columns}{a list containing "n" vectors of lenght 12 with all the conditions across columns in the plate}
}
\value{
"D" a data.table containing all the reads, times, and conditions
}
\description{
Function to import the reads from the plate reader at the Dr Summers Lab excluding the time variable
}
\author{
Santiago Caño-Muñiz
}
