% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_lavaan.R
\name{predict_lavaan}
\alias{predict_lavaan}
\title{predict_lavaan}
\usage{
predict_lavaan(fit, newdata = NULL)
}
\arguments{
\item{fit}{model fit with the lavaan package}

\item{newdata}{data to use for predictions}
}
\value{
data frame with predictions
}
\description{
predict_lavaan
}
\examples{
predict_lavaan(fit, newdata)
}
