% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.stars.R
\name{pval.stars}
\alias{pval.stars}
\title{Title: Function to turn p-values into stars*}
\usage{
It only takes a single value, so vectorising needs to be used here if \code{pval.stars} is to be applied
to a column or a vector with multiple elements.
}
\arguments{
\item{p}{A numeric vector.}

\item{symbol}{A string ("symbol") indicating the symbol to that will indicate significance.Default is `"*"`.}
}
\value{
It generates a vector of symbols (strings)
}
\description{
\code{pval.stars} converts the numeric values of a vector into specified sequences of symbols to signify p-value significance.
}
\examples{
(example_data <- data.frame(p.value = seq(0, 0.1, 0.02)))
example_data$stars <- sapply(example_data$p.value, pval.stars)
example_data


}
