% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_to_stars.R
\name{pval_to_start}
\alias{pval_to_start}
\title{P-values to start}
\usage{
pval_to_start(p, ns_na = FALSE)
}
\arguments{
\item{p}{numeric vector with probabilities from 0 to 1.}

\item{ns_na}{logical report non-significant values as NA.}
}
\description{
Transform a vector of probabilities into a significant starts
}
\author{
Santiago Caño-Muñiz
}
