% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range01.R
\name{range01}
\alias{range01}
\title{Range 0 to 1}
\usage{
range01(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Generic function to scale data in a 0-1 range using the maximum and minimum ratio
as described: \deqn{(x_i\ \ -\min\funcapply(x))/(\max\funcapply(x)-\min\funcapply(x)\ )}
}
\author{
Santiago Caño-Muñiz
}
