% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_funcs.r
\name{rescale_shp}
\alias{rescale_shp}
\title{rescale_shp}
\usage{
rescale_shp(shp, units = "m")
}
\arguments{
\item{shp}{object of class \code{sf}.}

\item{units}{distance units to rescale to (i.e. "m","km").}
}
\value{
a \code{sf} object
}
\description{
\code{rescale_shp} rescales a \code{sf} object with different distance units
 i.e. meters to km, feet to meters etc.
}
