% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsSampler.R
\name{Rgibbs}
\alias{Rgibbs}
\title{Gibbs sampler}
\usage{
Rgibbs(N, thin)
}
\arguments{
\item{N}{Number of iterations}

\item{thin}{sample}
}
\description{
Manual implementation of Gibbs Sampler in R
}
\author{
Santiago Caño-Muñiz
}
