% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_nearest.R
\name{round_nearest}
\alias{round_nearest}
\title{Round to the nearest number.}
\usage{
round_nearest(x, base = 5)
}
\arguments{
\item{x}{numeric: a numeric vector.}

\item{base}{integer: nearest integer to round round the number.}
}
\description{
Round to the nearest number.
}
\author{
Santiago Caño-Muñiz
}
