% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_in.R
\name{save_table}
\alias{save_table}
\title{save_table}
\usage{
save_table(
  what,
  file_name = "DF",
  where = getwd(),
  row_names = FALSE,
  fileEncoding = "latin1"
)
}
\arguments{
\item{what}{.}

\item{file_name}{.}

\item{where}{.}

\item{row_names}{.}

\item{fileEncoding}{.}
}
\value{
.
}
\description{
Funzione per ottenere una variabile data (formattata come stringa) a partire da un input di inserimento (xlsx).
}
