% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{scale_color_santi}
\alias{scale_color_santi}
\title{My palette for ggplot2.}
\usage{
scale_color_santi(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\description{
See link[MyFunctions]{get_mypalette}
}
\author{
Santiago Caño-Muñiz
}
