% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.R
\name{scatterplots}
\alias{scatterplots}
\title{scatterplots}
\usage{
scatterplots(
  x,
  y,
  ylim = c(),
  pch = c(),
  main = c(),
  xlab = c(),
  ylab = c(),
  data = c()
)
}
\arguments{
\item{x}{Numeric vector (variable x axis)}

\item{y}{Numeric vector (variable y axis)}

\item{ylim}{Y axis limits}

\item{pch}{pch as in plot}

\item{main}{Title of the plot}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{data}{Data frame}
}
\value{
formated scatter plots
}
\description{
scatterplots
}
