% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error.R
\name{se}
\alias{se}
\title{Standard error}
\usage{
se(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Calculates the standard error as: \deqn{(SE=\frac{\sigma }{\sqrt{n}})}
}
\author{
Santiago Caño-Muñiz
}
