% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat3ToMonocle3.R
\name{Seurat3ToMonocle3}
\alias{Seurat3ToMonocle3}
\title{convert different formats from different softwares or packages}
\usage{
Seurat3ToMonocle3(seuObj)
}
\arguments{
\item{seuObj}{the input seurat3 object}
}
\value{
an cell_data_set object
}
\description{
This is an package to convert formats among different softwares or packages
}
\details{
nothing
}
\examples{
pbmc_raw <- read.table(file = system.file('extdata', 'pbmc_raw.txt', package = 'Seurat'),as.is = TRUE)
pbmc_small <- CreateSeuratObject(counts = pbmc_raw)
pbmc_monocle3 <- Seurat3ToMonocle3(pbmc_small)

}
\author{
Feng Zhang
}
