% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.port.sort.R
\name{single.port.sort}
\alias{single.port.sort}
\title{Single portfolio sort}
\usage{
single.port.sort(
  FACTOR,
  WEIGHTING = "equal",
  HOLD,
  SubPort_No = 10,
  Output = "returns"
)
}
\arguments{
\item{FACTOR}{The cross-section of the sorting variable at time t-1.}

\item{WEIGHTING}{This variable specifies the weighting scheme.
For example, if you wish to do a value weighted portfolio sort, put the cross-sectional marcket cap here.
Otherwise, enter 'equal' as character, this will give you an equal weighted portfolio sort.}

\item{HOLD}{This is the asset returns at t.}

\item{SubPort_No}{A positive integer that specifies the number of portoflios to be formed.}

\item{Output}{A character indicating whether to give just returns for each portfolio or detailed ranking of assets in a list.
Possible values are: 'returns' and 'lists'.}
}
\value{
Either a vector of sub-portfolio returns or a list of detailed ranking of assets.

#' @export
}
\description{
This functions is the building block of single portfolio sort analyses.
It sorts and returns the results for one portfolio sort at a given time t.
By iterating this function over time, normal portoflio sort analysis can be done.
}
