% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.series.summary.stat.R
\name{single.series.summary.stat}
\alias{single.series.summary.stat}
\title{Summary statistics for a single series}
\usage{
single.series.summary.stat(
  single_series,
  scaler = 252,
  rounding_decimal = 2,
  Scaled = TRUE,
  SharpeR = TRUE,
  NWTstat = TRUE,
  Rounding = TRUE,
  Staring = TRUE
)
}
\arguments{
\item{single_series}{A series of data.}

\item{scaler}{An annualisation scaler applied to the stats computed from the raw data, if Scaled is TRUE.
The default value is 252.}

\item{rounding_decimal}{A positive integer indicating how many decimals to keep, if Rounding is TRUE.}

\item{Scaled}{A Boolean value specifying whether to scale the stats computed from the raw data.}

\item{SharpeR}{A Boolean value specifying whether to include Sharpe ratio.}

\item{NWTstat}{A Boolean value specifying whether to apply Newey-West correction.}

\item{Rounding}{A Boolean value specifying whether to round the results.}

\item{Staring}{A Boolean value specifying whether to give stars to MEAN that indicate statistical significance.}
}
\value{
A dataframe that contains summary statistics.
}
\description{
This function computes the summary statistics for a single series.
}
