% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothen.uvvis.R
\name{smoothen.uvvis}
\alias{smoothen.uvvis}
\title{Title: Function to smoothen curves for UV-vis data}
\usage{
x and y are strings with the respective column names.
}
\arguments{
\item{df}{A data frame.}

\item{spar}{Numerical value indicating amount of smoothing. See \code{\link{smooth.spline}}. Default is 0.5.}

\item{xIndex}{Numerical value indicating location of the x-axis column. Default is first column.}

\item{pivotLonger}{Logical value indicating whether the smoothen data frame should be pivoted for plotting. Default is F.}
}
\value{
A data frame with the smoothen curves.
}
\description{
\code{smoothen.uvvis} uses the smooth function \code{\link{smooth.spline}} to fit a cubic smoothing spline.
It then applies it to all the columns of a data frame, and provides a second data frame with the smoothen values.
Can also provide a pivoted data frame for plotting various curves.
}
\examples{
library(tidyverse)
# load example data
df <- read.csv("eg_rough_curve.csv")

# smoothen curves
smoothen_data <- smoothen.uvvis(df, xIndex = 1, pivotLonger = T)

# plotting smoothen data
ggplot(smoothen_data,
        aes(x = XYDATA,
            y = value,
            colour = name)) +
geom_line() +
theme_bw()

smoothen_data <- df[1:2]
smoothen_data$smooth <- smoothen.uvvis(df[1:2], pivotLonger = F)[[2]]
smoothen_data <- pivot_longer(smoothen_data, cols = !1)

# plotting side by side
ggplot(smoothen_data,
        aes(x = XYDATA,
            y = value,
            colour = name)) +
geom_line(show.legend = F) +
labs(x = "Wavelength (cm)", y = "Absorbance") +
facet_wrap(~name) +
theme_bw()
}
