% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{test_dates}
\alias{test_dates}
\title{A date format testing function}
\usage{
test_dates(mydates)
}
\arguments{
\item{mydates}{An array of dates to be tested}
}
\description{
A function for testing the format of a column of dates
}
\examples{
test_dates()
}
\keyword{date}
\keyword{format}

