% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_michelle.R
\name{theme_michelle}
\alias{theme_michelle}
\title{Michelle's 538 ggplot theme}
\usage{
theme_michelle(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{the base font size}

\item{base_family}{the base font family}
}
\description{
This function includes all of the changes I normally make to the 538 theme from the 'ggthemes' package
}
\examples{
data(iris)
ggplot(data=iris, aes(x=Sepal.Length, y=Sepal.Width)) +
  geom_point() +
  geom_point() +
  theme_my538()
}
