% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{theme for ggplot}
\usage{
theme_Publication(base_size = 14, base_family = "Calibri")
}
\arguments{
\item{base_size}{letter size}

\item{base_family}{leter type}
}
\description{
The theme_Publication generates like-publication graphs using the ggplot environment
}
\author{
Santiago Caño-Muñiz
}
