% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_scm.R
\name{theme_scm}
\alias{theme_scm}
\title{Stylish theme for ggplot}
\usage{
theme_scm(
  base_size = 15,
  base_family = "Calibri",
  x.text.angle = 0,
  border = FALSE,
  margin = TRUE,
  legend = c("top", "bottom", "left", "right", "none"),
  spacing = 0.5
)
}
\arguments{
\item{base_size}{letter size.}

\item{base_family}{leter type.}

\item{x.text.angle}{Inclination of X labels.}

\item{border}{Border of the box.}

\item{margin}{Add space between multi-facetted plots.}

\item{legend}{Location of the legend with a value "top", "bottom", "left", "right", "none".}

\item{spacing}{Sparation between facets in case there are.}
}
\description{
Helper function to create like-publication graphs using the ggplot environment.
}
\author{
Santiago Caño-Muñiz
}
