% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{thesis.theme}
\alias{thesis.theme}
\title{Custom theme for the thesis/publications}
\usage{
thesis.theme()
}
\value{
A ggplot2 theme object
}
\description{
Custom theme for the thesis/publications
}
\examples{
library(tidyverse)
library(ggprism)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  thesis.theme()

}
