% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{time.c}
\alias{time.c}
\title{Function to convert time from a string to a Posix object}
\usage{
\method{time}{c}(timechar, t_form = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{timechar}{A string or array of strings containing the time/date}

\item{t_form}{The format of the time (see strptime for format details) - default t_form="\%Y-\%m-\%d \%H:\%M:\%S"}
}
\description{
Function to convert time from a string to a Posix object
}
\examples{
time.c("2019-01-01 01:01:01", t_form="\%Y-\%m-\%d \%H:\%M:\%S")
}
