% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{timestamp}
\alias{timestamp}
\title{Current time stamp
This function when called returns the current time stamp. This can be generally used to append time stamp to file names when writing them multiple times without over writing.}
\usage{
timestamp(format = "\%y\%m\%d\%H\%M\%S")
}
\arguments{
\item{format}{Format of the timestamp to be returned}
}
\value{
timestamp
}
\description{
Current time stamp
This function when called returns the current time stamp. This can be generally used to append time stamp to file names when writing them multiple times without over writing.
}

