% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TRI.R
\name{TRI}
\alias{TRI}
\title{Compute a terrain ruggedness index followin sappington et al 2007}
\usage{
Terrain <- TRI(r= veg_mask,layer=veg_mask$elevation)
}
\arguments{
\item{r}{is the SpatialPixel(DataFrame)}

\item{layer}{the layer where elevation is stored (for instance r$elevation)}
}
\description{
it returns the sum of the absolute values of the elevation differences between a central grid cell and its 8 neighboring cells
}
\author{
Cyril Milleret
}

