% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv.plot.R
\name{uv.plot}
\alias{uv.plot}
\title{Title: Function for plotting uv-vis curves*}
\usage{
It takes a data frame pre-arranged in wavelength and absorbance columns and plots it. It has legend and saving options.
}
\arguments{
\item{df}{A data frame.}

\item{legend}{Logical vector indicating the inclusion of a legend in the plot.}

\item{legend_items}{A vector string with the legend elements. The legend must of equal length as the number of columns.}

\item{save}{Logical element for optional saving. Default is T.}

\item{savepath}{A string indicating path for saving.}

\item{smooth}{A logical vector indicating whether the curve should be smoothed, default is T. It uses \code{smoothen.uvvis} from this package to fit a cubic smoothing spline.}
}
\value{
A plot.
}
\description{
For plotting uv-vis data.
}
\examples{
library(tidyverse)
library(myfunctions)


}
