% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcanoplot}
\alias{volcanoplot}
\title{Make a volcanoplot of effect vs p-value and calculate the number of elements of the differently regulated groups}
\usage{
volcanoplot(x, pCol, effCol, minP, minEff, log10 = TRUE, plot = TRUE, ...)
}
\arguments{
\item{x}{Dataframe of matrix to be plotted}

\item{pCol}{Name of column contianing the p-values}

\item{effCol}{Name of column containing the effect sizes}

\item{minP}{p-value cutoff}

\item{minEff}{Effect size cutoff}

\item{log10}{Logical indicating whether to log transform the p-values}

\item{plot}{Logical indicating whether to plot or not}
}
\description{
This function takes a matrix or dataframe as input and upon specifying the correct columns output the number of elements in the differently regulated groups. Optionally, the volcanoplot is also plotted. A log10 transformation of the p-value column can also be applied
}
\examples{
volcanoplot(x, pCol="pVal", effCol = "log2FC", minP = 0.05, minEff = 1, log10 = TRUE, plot = TRUE)
}
