% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{weighted.m.sd}
\alias{weighted.m.sd}
\title{Function for weighted mean and standard deviation}
\usage{
weighted.m.sd(x, w, weightsaresd = TRUE)
}
\arguments{
\item{x}{Vector of data}

\item{w}{Vector of weights or standard deviations}

\item{weightsaresd}{If TRUE (default) then the weights are assumed to be the standard deviation
in the data, eg. a high value means a low weighting. If weightsaresd = FALSE the weights are
used directly, eg. a high weight is a high weighting. If weightsaresd = FALSE the returned
standard deviation in the weighted mean may be incorrect.}
}
\value{
A vector of three numbers
\itemize{
\item The weighted mean
\item The weighted standard deviation
\item The number of non-nan value-weight pairs
}
}
\description{
Function for weighted mean and standard deviation
}
