% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.intervals.return.cal.R
\name{x.intervals.return.cal}
\alias{x.intervals.return.cal}
\title{Return calculation}
\usage{
x.intervals.return.cal(
  timestamped_DATA,
  DATA_freq = "xts_minutely",
  x,
  PriceColumeName,
  style = "simple"
)
}
\arguments{
\item{timestamped_DATA}{An xts object that is a price series of which the index has class xts, Date, or yearmon.}

\item{DATA_freq}{A character specifying the underlying frequency of data.
For example, to comput half-hour returns from minutely stamped data, this value would be 'xts_minutely'.
possible values include 'xts_minutely', 'Date_daily', 'yearmon_monthly'.}

\item{x}{A positive integer indicating how long the interval, over which the return is computed, is.
For example, to comput half-hour returns from minutely stamped data, this value would be 30.}

\item{PriceColumeName}{The colunmn name of the price series.
If none, use integer 1.}

\item{style}{A character indicating whether to compute simple or log returns.
Possible values are 'simple' and 'log'.}
}
\value{
A xts series of returns.
}
\description{
This function computes the returns from a timestamped price series.
This function matches p_1 and p_2 by taking advantage of merge.xts, thus missing timestamps are easily dealt with.
}
