% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zComposite.R
\name{zComposite}
\alias{zComposite}
\title{zComposite}
\usage{
zComposite(
  x = c(),
  x_fu = c(),
  id = c(),
  sex_strata = FALSE,
  sex = "Sex",
  age = c("Age_I", "Age_II"),
  refs = NULL,
  df
)
}
\arguments{
\item{x}{Names of the variables to be included in the composite score}

\item{x_fu}{Names of the variables to be included in the composite score at follow up (it would calculate a z score of the change)}

\item{id}{Participant identifier variable}

\item{sex_strata}{Logical. Whether to stratify the composite score by sex}

\item{sex}{Name of the sex variable in the data frame.}

\item{age}{Name of the age variable in the dataset. Only used if reference values are provided in \code{refs}}

\item{refs}{Reference values as, for example, in https://doi.org/10.1016/j.atherosclerosis.2018.10.003 (important the reference values data frame presents the same colnames as the variable names in df)}

\item{df}{Data frame with variables to calculate the composite score}
}
\value{
Data frame with composite z score calculated
}
\description{
Calculates z composite scores of a set of variables, possibility to stratify by sex implemented
}
