% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero.norm.R
\name{zero.norm}
\alias{zero.norm}
\title{Title: Function for 0 to 1 curve normalisation*}
\usage{
It takes a data frame and an specified column (numeric position) and normalises according to a range of values in the x-axis.
For example, if there are multiple peaks in a spectra, then specifying the range for normalisation can exclude undesired peaks.
}
\arguments{
\item{df}{A data frame.}

\item{col}{A numeric vector indicating the position of the column in the data frame.}

\item{min_x}{Start of x-axis range, default = 300.}

\item{max_x}{End of x-axis range, default = 700.}
}
\value{
A numeric vector with normalised values.
}
\description{
\code{zero.norm} 0 to 1 curve normalisation.
}
