% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.portfolios.R
\name{assign.portfolios}
\alias{assign.portfolios}
\title{Assign Portfolios}
\usage{
assign.portfolios(
  data,
  break_var,
  sort_var,
  bps = NULL,
  n_portfolios = 10,
  P_prefix = "P",
  bp_prefix = "bp",
  show_bps = TRUE,
  bp_type = 2,
  form_type = 1
)
}
\arguments{
\item{data}{A tibble that (normally) contains at least two variables, one is used to compute breakpoints and the other is used to compute portfolio membership.
In the case that the breakpoints are computed from the sorting variable itself, the tibble could potentially contain only one variable.}

\item{break_var}{The variable that is used to compute breakpoints.}

\item{sort_var}{The variable that is used to compute portfolio membership.}

\item{bps}{A numerical vector specifying customised preakpoints.
It should contain only the breakpoints not the end values.
For example, for a tercile sort, this value should be c(0.2, 0.8), not c(0, 0.2, 0.8, 1).}

\item{n_portfolios}{A possitive number.
This is the umber of portfolios to which assets are evenly assigned.
Breakpoints are computed evenly across the range in each cross-section.}

\item{P_prefix}{The prefix to be placed in front of portfolio numbers.
This will be part of the column names for the portfolio membership columns.
The default is 'P', which gives column names P1, P2, P3, ...}

\item{bp_prefix}{Similar to P_prefix, this is the prefix to be placed in front of breakpoint numbers.
This will be part of the column names for the breakpoint columns.
The default is 'bp', which gives column names bp1, bp2, ...}

\item{show_bps}{A Boolean value indicating whether to retain breakpoints in the output dataframe.
When FALSE, the breakpoints will not be shown and the argument bp_prefix will be ignored.}

\item{bp_type}{An integer between 1 and 9 selecting one of the nine quantile algorithms used by quantile(), see which for details.
The default value is 2 which is the default algorithm used by SAS.}

\item{form_type}{An integer between 1 and 4 selecting one of the four portfolio assignment methods:
\itemize{
\item 1 selects the portfolio formation method detailed in Bali, Engle, and Murray (2016). It assigns a value to Portfolio 1 if it is in \link{bp1, bp2}, i.e., values at the boundaries are assigned to more than one portfolio.
\item 2 selects the method that assigns a value to Portfolio 1 if it is in [bp1, bp2). No values will be assigned to more than one portfolio.
\item 3 selects the method that assigns a value to Portfolio 1 if it is in (bp1, bp2]. No values will be assigned to more than one portfolio.
}

[bp1, bp2). No values will be assigned to more than one portfolio.
\itemize{
\item 3 selects the method that assigns a value to Portfolio 1 if it is in (bp1, bp2]: R:bp1,\%20bp2).\%20No\%20values\%20will\%20be\%20assigned\%20to\%20more\%20than\%20one\%20portfolio.\%0A\%20*\%203\%20selects\%20the\%20method\%20that\%20assigns\%20a\%20value\%20to\%20Portfolio\%201\%20if\%20it\%20is\%20in\%20(bp1,\%20bp2
}}
}
\value{
A tibble that contains two types of columns (variables), one is the portfolio membership columns, the other is the breakpoint columns.
Each portfolio membership column gives Boolean values indicating whether an asset belongs to that portfolio.
Each breakpint column gives the breakpoint used for computing membership, this will be one column more than the portfolio memebership columns.
}
\description{
This function assigns each of the assets to a portfolio in the cross-section.
By defult, this function implements the portfolio sorting method as in Bali et al. (2016).
That is, it is possible for an asset to be assigned to more than one portfolios in a given cross-section.
}
