% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_paramter_bootstrap_function.R
\name{boot_param_CI}
\alias{boot_param_CI}
\title{Bootstrapping confidence intervals for lme4 and glmmTMB}
\usage{
boot_param_CI(nsim, model, data)
}
\arguments{
\item{nsim}{number of iteration for the bootstrap}

\item{model}{lme4 (merMod) or glmmTMB model object}

\item{data}{dataframe used in the model}
}
\value{
returns model parameters for all bootstrap iterations and 95\% confidence interval plots and dataframe
}
\description{
Bootstrapping confidence intervals for lme4 and glmmTMB
}
