% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotRegression.R
\name{ggplotRegression}
\alias{ggplotRegression}
\title{Linear model equations}
\usage{
ggplotRegression(m)
}
\arguments{
\item{base_size}{letter size}

\item{base_family}{leter type}
}
\description{
Plot the equation of simple linear models generated with lm.
}
\author{
Santiago Caño-Muñiz
}
