% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted_descriptions.R
\name{give_categorical_crosstable}
\alias{give_categorical_crosstable}
\title{give_categorical_crosstable}
\usage{
give_categorical_crosstable(
  x = NA,
  y = NA,
  name_x = "",
  max_missed = 0.1,
  void_string = "-",
  list_marker = "-",
  alpha_value = 0.05,
  multiple_alphas = c(0.05, 0.01, 0.001)
)
}
\arguments{
\item{x}{Un vettore di fattori, convertito come tale se numerico.}

\item{y}{Un vettore di fattori, convertito come tale se numerico, con il gruppo di interesse.}

\item{name_x}{Stringa da usare come nome della variabile (altrimenti è utilizzata l'etichetta, se presente).}

\item{max_missed}{Percentuale massima ammessa di omissioni.}

\item{void_string}{Stringa da usare se il valore non c'è o non è calcolabile.}

\item{list_marker}{Marcatore di punto elenco (per la lista dei livelli).}

\item{alpha_value}{Valore di alpha per la significatività del test.}

\item{multiple_alphas}{Valori di alpha per gli asterischi sulla significatività del test.}
}
\value{
Un data.frame con i risultati descrittivi (una colonna per livello della variabile con i gruppi).
}
\description{
Funzione per descrivere una variabile categoriale e il suo incrocio con un'altra di cui interessa la divisione in gruppi (con una tabella relativa a tutti i livelli possibili).
}
