% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_basic_tests.R
\name{give_categorical_test}
\alias{give_categorical_test}
\title{give_categorical_test}
\usage{
give_categorical_test(x = NA, y = NA, void_string = "-")
}
\arguments{
\item{x}{Vettore di fattori.}

\item{y}{Vettore di fattori.}

\item{void_string}{Stringa da usare se il valore non c'è o non è calcolabile.}
}
\value{
Un vettore con il risultato del test e con il valore di p risultante.
}
\description{
Test per categorie (Chi quadro o test di Fisher).
}
