% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_basic_tests.R
\name{give_chisquare}
\alias{give_chisquare}
\title{give_chisquare}
\usage{
give_chisquare(
  x = NA,
  y = NA,
  void_string = "-",
  alpha_value = 0.05,
  multiple_alphas = c(0.05, 0.01, 0.001)
)
}
\arguments{
\item{x}{Vettore di fattori.}

\item{y}{Vettore di fattori.}

\item{void_string}{Stringa da usare se il valore non c'è o non è calcolabile.}

\item{alpha_value}{Valore di alpha per la significatività del test.}

\item{multiple_alphas}{Valori di alpha per gli asterischi sulla significatività del test.}
}
\value{
Un vettore con il risultato del test, con il valore di p risultante e con la comparazione fra i gruppi.
}
\description{
Test del Chi quadrato.
}
