% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func,lag_cal.R
\name{lag.cal}
\alias{lag.cal}
\title{Computing Optimal Lags for Newey-West tests}
\usage{
\method{lag}{cal}(x)
}
\arguments{
\item{x}{A series of data.}
}
\value{
The optimal lags for Newey-West tests.
}
\description{
This function computes the optimal lags for Newey-West tests.
}
