% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_coda.R
\name{lm_coda}
\alias{lm_coda}
\title{Basic function for cross-sectional linear regression CoDA models}
\usage{
lm_coda(data, compo, outcome, covariates = c(), scale = FALSE)
}
\arguments{
\item{data}{data frame or matrix with variables of interest}

\item{compo}{String with colnames of composition variables or matrix/dataframe with these variables}

\item{outcome}{String with colname of outcome or numeric vector with the outcome}

\item{covariates}{String with names of covariates (optional) or matrix/dataframe with these variables}

\item{scale}{Logica. Whether to standardize results (default = FALSE)}
}
\value{
Linear regression model (lm class)
}
\description{
Basic function for cross-sectional linear regression CoDA models
}
