% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ggtheme_m.R
\name{my_ggtheme_m}
\alias{my_ggtheme_m}
\title{My personal map theme}
\usage{
my_ggtheme_m(
  map_type = "global",
  ax_tx_s = 14,
  axx_tx_ang = 0,
  axy_tx_ang = 0,
  ax_tl_s = 15,
  leg_pos = "bottom",
  leg_aline = 0,
  leg_tl_s = 16,
  leg_tx_s = 14,
  leg_width = 1,
  hjust = 0,
  facet_tx_s = 10
)
}
\value{
a ggplot theme
}
\description{
This function creates my own ggplot theme to 
standarize multiple map in one document
}
