% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ggtheme_p.R
\name{my_ggtheme_p}
\alias{my_ggtheme_p}
\title{My personal plot theme}
\usage{
my_ggtheme_p()
}
\value{
a ggplot theme
}
\description{
This function creates my own ggplot theme to 
standarize multiple plots in one document
}
