% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ggtheme_p.R
\name{my_ggtheme_p}
\alias{my_ggtheme_p}
\title{My personal plot theme}
\usage{
my_ggtheme_p(
  ax_tx_s = 9,
  axx_tx_ang = 0,
  axy_tx_ang = 0,
  ax_tl_s = 10,
  leg_pos = "top",
  leg_aline = 0.5,
  leg_tl_s = 10,
  leg_tx_s = 9,
  x_hjust = 0,
  x_vjust = 0,
  y_hjust = 0,
  y_vjust = 0,
  facet_tx_s = 12
)
}
\value{
a ggplot theme
}
\description{
This function creates my own ggplot theme to 
standarize multiple plots in one document
}
