% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_paths.R
\name{my_path}
\alias{my_path}
\title{Load and installs packages needed}
\usage{
my_path(Path_Type, Project = NA, Extra_Path = "", Name = "",
  Repo = TRUE, Read = FALSE, header = TRUE)
}
\arguments{
\item{Path_Type}{Type of path needed "D" for Data, "R", for Result and "G" for generic. You need those paths to be load in the global environment}

\item{Project}{I normally call my data folder the same as the project so it automatically sets to the project name. Alternatives available.}

\item{Extra_Path}{Any sub-path to add WITHIN the folders. Note, if it is before, include it on the "Folder" parameter -e.g. Project/SubFolder -}

\item{Name}{Dataset name in case you want the option to load data}

\item{Repo}{Normally set to jepa88 but if connection to Drobo is availbale use Repo = "FALSE"}

\item{Read}{Default Read = FALSE. Read = FALSE will only return a path wile Read = TURE will read a dataset}
}
\value{
Paths to save and load data within a project
}
\description{
This function loads all necessary packages. If the
package is not installed, it will install it first
and the it will load the recently installed package
}
