% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_paths.R
\name{my_path}
\alias{my_path}
\title{Load and installs packages needed}
\usage{
my_path(
  path_type,
  extra_path = "",
  name = "",
  list_files = "NA",
  repo = TRUE,
  read = FALSE,
  header = TRUE,
  project = TRUE,
  system = NA
)
}
\arguments{
\item{path_type}{Type of path needed "D" for Data, "R", for Result and "G" for generic. You need those paths to be load in the global environment}

\item{extra_path}{Any sub-path to add WITHIN the folders. Note, if it is before, include it on the "Folder" parameter -e.g. project/SubFolder -}

\item{name}{Dataset name in case you want the option to load data}

\item{repo}{Normally set to jepa88 but if connection to Drobo is available use repo = "FALSE"}

\item{read}{Default read = FALSE. read = FALSE will only return a path wile read = TURE will read a dataset}

\item{project}{I normally call my data folder the same as the project so it automatically sets to the project name. Alternatives available.}

\item{system}{If you want to override the automatic system identification}
}
\value{
Paths to save and load data within a project
}
\description{
This function loads all necessary packages. If the
package is not installed, it will install it first
and the it will load the recently installed package
}
