% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_sf.R
\name{my_sf}
\alias{my_sf}
\title{Loads a specific shapefile}
\usage{
my_sf(sf_name, crs = "NA", simple = "NA", system = NA)
}
\arguments{
\item{sf_name}{Name of shapefile to load. Current options FAO, SAU, MEOW, RFMO}

\item{crs}{The need crs of shapefile}

\item{simple}{if simple = NA then is fine}

\item{system}{If you want to override the automatic system identification.}
}
\value{
the requested shapefile of choice
}
\description{
This function simply loads a shapefile. I just created it because
I have frquently uised sf and is annoying to be coping + pasting
every single time
}
