% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newey.west.one.sample.t.R
\name{Newey.west.one.sample.t}
\alias{Newey.west.one.sample.t}
\title{Newey-West One Sample t-stat}
\usage{
Newey.west.one.sample.t(dat)
}
\arguments{
\item{dat}{A series of data}
}
\value{
The Newey-West t-stat
}
\description{
This function computes the Newey-West One Sample t-stat for a single series.
The optimal lag is computed by the function lag.cal.
}
