% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_overlap_index.R
\name{niche_overlap_index}
\alias{niche_overlap_index}
\title{Index of habitat niche overalp}
\usage{
niche_overlap_index(X, kernel = "gaussian", bw = "nrd0")
}
\arguments{
\item{X}{a kselect (adehabitatHS) object}

\item{kernel}{a character string giving the smoothing kernel to be used. Default is "gaussian". See \code{density()}}

\item{bw}{the smoothing bandwidth to be used to compute the kernel density. Default is "nrd0". See \code{density()}}
}
\value{
A list of paired matrices with an overlap index between each individuals calculated on each axis and a Weighted (based on eigen values of the kselect) overlap index calculated on all axes
}
\description{
Derive an index (0: complete separation, 1: complete overlap) from a kselect object to estimate a degree of habitat niche overlap between species/indviduals
}
\details{
Uses scores obtained for each individual on each dimension of the ecological space of the Kselect and compute a non-parametric kernel estimation. 
Area of overlap between distribution of the two indviduals is calculated. Eigen values of the kslect are then used to calculated an overlap weihgted index.
 \deqn{O_{ij} = \int_{b}^{a} \left [ f(x) - g(x) \right ]dx}
 \deqn{\bar{O_{ij}} =  \frac{\sum_{e}^{n} O_{ij_{e}} \lambda _{e}}{\sum_{e}^{n} \lambda _{e}}}
i and j are the two individuals to be compared, f(x) and f(x) the two probability functions of indviduals i and j and e each of the axes.
}
\examples{
library(adehabitatHS)
data(puechabonsp)

locs <- puechabonsp$relocs
map <- puechabonsp$map

## compute the home range of animals (e.g. using the minimum convex polygon)
 
pc <- mcp(locs[,"Name"])
hr <- hr.rast(pc, map)

## Compute the number of relocation in each pixel of the map
cp <- count.points(locs[,"Name"], map)

## prepares the data for the kselect analysis
x <- prepksel(map, hr, cp)
tab <- x$t

## We call a new graphic window
## A K-select analysis
acp <- dudi.pca(tab, scannf = FALSE, nf = 4)
kn <- kselect(acp, x$factor, x$weight,
              scannf = FALSE, nf = 4)

# compute habitat niche overlap 
niche_overlap_index(kn, kernel = "gaussian", bw = "nrd0")
}
\author{
Cyril Milleret
}

