% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reallocationPlot.R
\name{reallocationPlot}
\alias{reallocationPlot}
\title{reallocationPlot}
\usage{
reallocationPlot(
  data = c(),
  comp = c(),
  comp_fup = NULL,
  comp.names = c(),
  outcome = c(),
  covs = c(),
  comparisons = c("one-v-one", "prop-realloc")[1],
  increase = c(),
  decrease = c(),
  xlab = c(),
  ylab = c(),
  xlim = c(),
  ylim = c(),
  xaxis.by = 5,
  ribbon = TRUE,
  font = "Times New Roman",
  main = NULL,
  col = "black",
  alpha = 0.2
)
}
\arguments{
\item{data}{Data frame with the data}

\item{comp}{String with the name of the composition variables in dataf}

\item{comp_fup}{String with the name of the follow-up composition variables in dataf (optional)}

\item{comp.names}{Names of the composition variables to appear in the plot (it will use \code{comp} if not provided)}

\item{outcome}{String with the name of the outcome in dataf}

\item{covs}{String with the name of the covariates}

\item{comparisons}{Two choices ("one-v-one" or "prop-realloc")}

\item{increase}{What variable in the composition should increase in the visualization (replacement)}

\item{decrease}{What variable/s in the composition should decrease in the visualization (replaced)}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{xlim}{Limits of the X axis (time reallocation limits)}

\item{ylim}{Limits for the Y axis (autogenerated if not provided)}

\item{xaxis.by}{X axis ticks at every \code{xaxis.by} minutes}

\item{ribbon}{Logical. Whether to plot confidence intervals of the predictions}

\item{font}{Font to use (default = "Times New Roman")}

\item{main}{Title of the plot}

\item{col}{Color for lines and confidence intervals}

\item{alpha}{Alpha for transparency in confidence intervals}
}
\value{
reallocation plot with predictions
}
\description{
Visualization of predictions from CoDA model
}
\details{
If follow-up composition is provided, it will run prospective and change models/visualizations
}
