% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{scale_color_santi}
\alias{scale_color_santi}
\alias{scale_fill_santi}
\title{My palette wrapper for ggplot2.}
\usage{
scale_color_santi(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_fill_santi(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{`character`: Name of the colour set. Currently: "contrast", "main", "cool", "hot", "mixed" and "grey".}

\item{discrete}{`logical`: selects palette for continuous or discrete data}

\item{reverse}{`logical`: Should the order of the colors be reversed?}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:scale_color_gradientn]{ggplot2::scale_color_gradientn}}
  \describe{
    \item{\code{space}}{colour space in which to calculate gradient. Must be "Lab" -
other values are deprecated.}
    \item{\code{na.value}}{Colour to use for missing values}
    \item{\code{guide}}{Type of legend. Use \code{"colourbar"} for continuous
colour bar, or \code{"legend"} for discrete colour legend.}
    \item{\code{aesthetics}}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}
    \item{\code{colours}}{Vector of colours to use for n-colour gradient.}
    \item{\code{colors}}{Vector of colours to use for n-colour gradient.}
    \item{\code{values}}{if colours should not be evenly positioned along the gradient
this vector gives the position (between 0 and 1) for each colour in the
\code{colours} vector. See \code{\link[scales:rescale]{rescale()}} for a convenience function
to map an arbitrary range to between 0 and 1.}
  }}
}
\description{
See link[MyFunctions]{get_mypalette}
}
\author{
Santiago Caño-Muñiz
}
