% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific_axis.R
\name{scientific_axis}
\alias{scientific_axis}
\title{Scientific axis}
\usage{
scientific_axis(x)
}
\arguments{
\item{x}{numeric vector to be transform}
}
\description{
Transform a numeric vector into a character expression in scietific format
}
\author{
Santiago Caño-Muñiz
}
