% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{theme for ggplot}
\usage{
theme_Publication(
  base_size = 14,
  base_family = "Arial",
  x.text.angle = 0,
  border = FALSE,
  margin = TRUE,
  legend = c("top", "bottom", "left", "right", "none"),
  spacing = 0.5
)
}
\arguments{
\item{base_size}{letter size}

\item{base_family}{leter type}

\item{x.text.angle}{Inclination of X labels}

\item{border}{Border of the box}

\item{legend}{Location of the legend with a value "top", "bottom", "left", "right", "none"}

\item{spacing}{Sparation between facets in case there are}
}
\description{
Theme inspired in the Adnres publication style.
}
\author{
Santiago Caño-Muñiz
}
