% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TRI.R
\name{TRI}
\alias{TRI}
\title{Compute a terrain ruggedness index followin sappington et al 2007}
\usage{
Terrain <- TRI(r1= veg_mask,res=3)
}
\arguments{
\item{r1}{is the raster layer to compute TRI from}

\item{gri}{is the resolution of the grid in which TRI should be calculated. grid=3 will look like this :
1,1,1,
1,0,1,
1.1,1  where 0 is the focal cell where the index is calculated and 1 the cells use to calculated the relative index}
}
\description{
it returns a raster layer with the sum of the absolute values of the elevation differences between a central grid cell and its 8 neighboring cells
}
\author{
Cyril Milleret
}

