% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv.plot.R
\name{uv.plot}
\alias{uv.plot}
\title{Title: Function for plotting uv-vis curves*}
\usage{
It takes a data frame pre-arranged in wavelenth and absorbance columns and plots it. It has legend and saving options.
}
\arguments{
\item{df}{A data frame.}

\item{legend}{A vector string with the legend elements. The legend must of equal length as the number of columns.}

\item{save}{Logical element for optinal saving. Default is T.}

\item{savepath}{A string indicating path for saving.}
}
\value{
A plot.
}
\description{
\code{uv-plot} Plotting uv-vis absorbance data.
}
